@@ -1,7 +1,7 @@
 'use strict';
 
 // Load modules
-const EventEmitter = require('events').EventEmitter;
+import { EventEmitter } from 'events';
 
 /**
  *
@@ -117,6 +117,4 @@ const BotInterface = class extends EventEmitter {
   }
 };
 
-module.exports = {
-  BotInterface,
-};
+export { BotInterface };
