@@ -1,27 +1,54 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
-const url = require('url');
+import * as lodash from 'lodash-es';
+import { parse } from 'url';
 
-const handlebars = require('handlebars');
-const moment = require('moment');
-const logger = require('../utils/logger');
-const { BotInterface } = require('./bot-interface');
-const { Connector } = require('./connector');
-const { CommandFactory } = require('../command/command-factory');
-const EventEmitter = require('events').EventEmitter;
-const { EventStore } = require('./event-store');
-const { Hook } = require('./hook');
-const HttpsProxyAgent = require('https-proxy-agent');
-const { MockConnector } = require('./mock-connector');
-const messageParser = require('../command/message');
-const storage = require('../storage/storage');
-const responseHandler = require('./response-handler');
-const channelsApi = require('../slack-api/channels');
-const usersApi = require('../slack-api/users');
-const userGroupsApi = require('../slack-api/user-groups');
+import Handlebars from 'handlebars';
+import * as moment from 'moment';
+import logger from '../utils/logger.js';
+import { BotInterface } from './bot-interface.js';
+import { Connector } from './connector.js';
+import { CommandFactory } from '../command/command-factory.js';
+import { EventEmitter } from 'events';
+import { EventStore } from './event-store.js';
+import { Hook } from './hook.js';
+import HttpsProxyAgent from 'https-proxy-agent';
+import { MockConnector } from './mock-connector.js';
+import { parse as _parse } from '../command/message.js';
+import { Storage } from '../storage/storage.js';
+import {
+  isDirectMessage,
+  isPrivateMessage,
+  generateErrorTemplate,
+  generateBotResponseTemplate,
+} from './response-handler.js';
+import { getChannelsList } from '../slack-api/channels.js';
+import { getUsersList } from '../slack-api/users.js';
+import { getUserGroupsList } from '../slack-api/user-groups.js';
 
+const {
+  map,
+  get,
+  toUpper,
+  nth,
+  reduce,
+  isFunction,
+  isArray,
+  forEach,
+  isBoolean,
+  merge,
+  extend,
+  toLower,
+  endsWith,
+  replace,
+  compact,
+  repeat,
+  split,
+  find,
+  remove,
+} = lodash;
+const { unix } = moment;
 /**
  *
  * Represents the state and events of a bot.
@@ -39,18 +66,18 @@ const Bot = class {
     this.hook = {};
     this.eventEmitter = new EventEmitter();
     this.interfaceEventEmitter = new EventEmitter();
-    this.botMessageParser = messageParser.parse(
-      _.map(_.get(this.config, 'botCommand'), (command, key) => {
+    this.botMessageParser = _parse(
+      map(get(this.config, 'botCommand'), (command, key) => {
         return {
-          command: _.toUpper(key),
-          alias: command.alias ? (command.alias || []).map(_.toUpper) : [],
+          command: toUpper(key),
+          alias: command.alias ? (command.alias || []).map(toUpper) : [],
         };
       })
     );
 
     this.setupEvents();
     this.setupIntefaceEvents();
-    handlebars.registerHelper(this.registerHandlebarsHelpers());
+    Handlebars.registerHelper(this.registerHandlebarsHelpers());
   }
 
   /**
@@ -59,11 +86,11 @@ const Bot = class {
    * @return {Object} Promise object resolves to @link bot-interface.js
    */
   init() {
-    if (_.get(this.config, 'mock')) {
+    if (get(this.config, 'mock')) {
       logger.debug('Connecting for bot in mock');
       this.connectionManager = new MockConnector(this.config.botToken, {
         socketEventEmitter: this.eventEmitter,
-        mock: _.get(this.config, 'mock'),
+        mock: get(this.config, 'mock'),
       });
 
       this.connectionManager.connect();
@@ -77,8 +104,8 @@ const Bot = class {
 
       this.getAllTeamData(this.config.slackApi)
         .then((values) => {
-          this.getSlackData()['channels'] = _.nth(values, 0);
-          this.getSlackData()['members'] = _.nth(values, 1);
+          this.getSlackData()['channels'] = nth(values, 0);
+          this.getSlackData()['members'] = nth(values, 1);
 
           return this.connectionManager.connect();
         })
@@ -197,17 +224,17 @@ const Bot = class {
       message
     );
     const isMessageForBot =
-      _.toUpper(this.getBotName()) === parsedMessage.message.commandPrefix ||
-      _.toUpper(this.getId()) === parsedMessage.message.commandPrefix ||
-      _.toUpper(this.getEId()) === parsedMessage.message.commandPrefix;
+      toUpper(this.getBotName()) === parsedMessage.message.commandPrefix ||
+      toUpper(this.getId()) === parsedMessage.message.commandPrefix ||
+      toUpper(this.getEId()) === parsedMessage.message.commandPrefix;
     const isSelf =
-      parsedMessage.user === _.toUpper(this.getId()) ||
-      parsedMessage.user === _.toUpper(this.getEId());
+      parsedMessage.user === toUpper(this.getId()) ||
+      parsedMessage.user === toUpper(this.getEId());
 
     if (
       this.config.blockDirectMessage &&
-      (responseHandler.isDirectMessage(message) ||
-        (responseHandler.isPrivateMessage(message) && isMessageForBot))
+      (isDirectMessage(message) ||
+        (isPrivateMessage(message) && isMessageForBot))
     ) {
       logger.info('processed message ', parsedMessage);
 
@@ -217,10 +244,7 @@ const Bot = class {
       );
     }
 
-    if (
-      !isSelf &&
-      (responseHandler.isDirectMessage(message) || isMessageForBot)
-    ) {
+    if (!isSelf && (isDirectMessage(message) || isMessageForBot)) {
       logger.info('processed message ', parsedMessage);
 
       return this.commandFactory.handleMessage(parsedMessage).catch((err) => {
@@ -330,14 +354,13 @@ const Bot = class {
     if (this.store) {
       this.botInterface.emit('connect');
     } else {
-      storage
-        .getEvents(['events', 'schedule'])
+      Storage.getEvents(['events', 'schedule'])
         .then((events) => {
           this.store = new EventStore({
-            eventStore: _.reduce(
+            eventStore: reduce(
               events,
               (result, value, key) => {
-                result[key] = _.get(value, this.getBotName());
+                result[key] = get(value, this.getBotName());
 
                 return result;
               },
@@ -390,12 +413,12 @@ const Bot = class {
    * @callback Function called after socket message sent.
    */
   dispatchMessage(options, callback) {
-    callback = _.isFunction(callback) ? callback : undefined;
-    options.channels = _.isArray(options.channels)
+    callback = isFunction(callback) ? callback : undefined;
+    options.channels = isArray(options.channels)
       ? options.channels
       : [options.channels || options.channel];
 
-    _.forEach(options.channels, (channel) => {
+    forEach(options.channels, (channel) => {
       const message = {
         id: new Date().getTime().toString(),
         type: options.type || 'message',
@@ -430,7 +453,7 @@ const Bot = class {
    * @return {object} Promise resolves the rendered error message.
    */
   handleErrorMessage(botName, context) {
-    const renderedData = responseHandler.generateErrorTemplate(
+    const renderedData = generateErrorTemplate(
       botName,
       this.config.botCommand,
       context
@@ -456,8 +479,8 @@ const Bot = class {
   handleBotMessages(parsedMessage, botDirectMessageError) {
     let renderedMessage;
     try {
-      if (!_.isBoolean(botDirectMessageError)) {
-        const error = handlebars.compile(botDirectMessageError)({
+      if (!isBoolean(botDirectMessageError)) {
+        const error = Handlebars.compile(botDirectMessageError)({
           user: `<@${parsedMessage.user}>`,
         });
         renderedMessage = botDirectMessageError ? error : '';
@@ -467,7 +490,7 @@ const Bot = class {
       renderedMessage = '';
     }
 
-    const renderedData = responseHandler.generateBotResponseTemplate({
+    const renderedData = generateBotResponseTemplate({
       botDirectMessageError: Boolean(botDirectMessageError),
       message: renderedMessage ? renderedMessage : false,
     });
@@ -487,7 +510,7 @@ const Bot = class {
    * @return {string} Bot id.
    */
   getEId() {
-    const socket = _.get(this, 'connectionManager.socket');
+    const socket = get(this, 'connectionManager.socket');
     return socket ? socket.getEId() : undefined;
   }
 
@@ -528,7 +551,7 @@ const Bot = class {
    * @return {string} Bot id.
    */
   getId() {
-    const socket = _.get(this, 'connectionManager.socket');
+    const socket = get(this, 'connectionManager.socket');
     return socket ? socket.getId() : undefined;
   }
 
@@ -538,7 +561,7 @@ const Bot = class {
    * @return {string} Bot name.
    */
   getBotName() {
-    const socket = _.get(this, 'connectionManager.socket');
+    const socket = get(this, 'connectionManager.socket');
     return socket ? socket.getBotName() : undefined;
   }
 
@@ -548,7 +571,7 @@ const Bot = class {
    * @return {object} Slack data.
    */
   getSlackData() {
-    const socket = _.get(this, 'connectionManager.socket');
+    const socket = get(this, 'connectionManager.socket');
     return socket ? socket.getSlackData() : {};
   }
 
@@ -568,7 +591,7 @@ const Bot = class {
    * @return {object} Promise resolves to success/failure.
    */
   injectMessage(messageObj) {
-    const message = _.merge(
+    const message = merge(
       {},
       {
         id: '',
@@ -618,7 +641,7 @@ const Bot = class {
    */
   getSocketAgent() {
     if (!this.socketAgent && this.proxy && this.proxy.url) {
-      const opts = url.parse(this.proxy.url);
+      const opts = parse(this.proxy.url);
       opts.secureEndpoint = opts.protocol ? opts.protocol == 'wss:' : false;
       this.socketAgent = new HttpsProxyAgent(opts);
     }
@@ -633,7 +656,7 @@ const Bot = class {
    */
   getHttpAgent() {
     if (!this.httpAgent && this.proxy && this.proxy.url) {
-      const opts = url.parse(this.proxy.url);
+      const opts = parse(this.proxy.url);
       opts.secureEndpoint = opts.protocol ? opts.protocol == 'https:' : false;
       this.httpAgent = new HttpsProxyAgent(opts);
     }
@@ -648,14 +671,14 @@ const Bot = class {
    * @return {array} List of slack users. Look at slack docs for user model.
    */
   fetchTeamUsers(config) {
-    if (_.get(config, 'exclude', false)) {
+    if (get(config, 'exclude', false)) {
       logger.info('Excluding user list call');
       return {};
     }
 
     logger.info('Fetching user list from slack');
 
-    return usersApi.getUsersList(config);
+    return getUsersList(config);
   }
 
   /**
@@ -666,7 +689,7 @@ const Bot = class {
    * Look at slack docs for channel model.
    */
   fetchTeamChannels(config) {
-    if (_.get(config, 'exclude', false)) {
+    if (get(config, 'exclude', false)) {
       logger.info('Excluding channel list call');
 
       return {};
@@ -674,7 +697,7 @@ const Bot = class {
 
     logger.info('Fetching channel list from slack');
 
-    return channelsApi.getChannelsList(config);
+    return getChannelsList(config);
   }
 
   /**
@@ -685,7 +708,7 @@ const Bot = class {
    * Look at slack docs for user group model.
    */
   fetchTeamUserGroups(config) {
-    if (_.get(config, 'exclude', false)) {
+    if (get(config, 'exclude', false)) {
       logger.info('Excluding channel list call');
 
       return {};
@@ -693,7 +716,7 @@ const Bot = class {
 
     logger.info('Fetching user group list from slack');
 
-    return userGroupsApi.getUserGroupsList(config);
+    return getUserGroupsList(config);
   }
 
   /**
@@ -709,8 +732,8 @@ const Bot = class {
     };
 
     return Promise.all([
-      this.fetchTeamChannels(_.extend(_.get(config, 'channel'), options)),
-      this.fetchTeamUsers(_.extend(_.get(config, 'user'), options)),
+      this.fetchTeamChannels(extend(get(config, 'channel'), options)),
+      this.fetchTeamUsers(extend(get(config, 'user'), options)),
     ]);
   }
 
@@ -743,58 +766,56 @@ const Bot = class {
       },
       toLowerCase: (context) => {
         if (context) {
-          return _.toLower(context);
+          return toLower(context);
         }
       },
       formatUnixTime: (context) => {
-        const time = _.get(context, 'hash.value', 0);
-        const offset = Number(_.get(context, 'hash.offset', 0));
-        const format = _.get(context, 'hash.format', 'YYYY-MM-DD HH:mm');
+        const time = get(context, 'hash.value', 0);
+        const offset = Number(get(context, 'hash.offset', 0));
+        const format = get(context, 'hash.format', 'YYYY-MM-DD HH:mm');
 
-        return moment.unix(time).utcOffset(offset).format(format);
+        return unix(time).utcOffset(offset).format(format);
       },
       joinArray: (context, options) => {
-        const separator = _.get(options, 'hash.separator', ' ');
-        const prefix = _.get(options, 'hash.prefix', '');
-        const suffix = _.get(options, 'hash.suffix', '');
+        const separator = get(options, 'hash.separator', ' ');
+        const prefix = get(options, 'hash.prefix', '');
+        const suffix = get(options, 'hash.suffix', '');
         const arrayVal = context || [];
 
-        if (!_.isArray(arrayVal)) {
+        if (!isArray(arrayVal)) {
           return `${arrayVal}`;
         }
 
-        return (
-          prefix + (_.map(arrayVal, _.toLower) || []).join(separator) + suffix
-        );
+        return prefix + (map(arrayVal, toLower) || []).join(separator) + suffix;
       },
       arrayItem: (context) => {
-        const index = _.get(context, 'hash.index', 0);
-        const arrayVal = _.get(context, 'hash.value', 0);
+        const index = get(context, 'hash.index', 0);
+        const arrayVal = get(context, 'hash.value', 0);
 
-        if (!_.isArray(arrayVal)) {
+        if (!isArray(arrayVal)) {
           return arrayVal;
         }
 
-        return _.nth(arrayVal, index);
+        return nth(arrayVal, index);
       },
       safeString: (context) => {
-        return new handlebars.SafeString(context);
+        return new Handlebars.SafeString(context);
       },
       toPlural: (context) => {
-        const length = _.get(context, 'hash.len', 0);
-        const word = _.get(context, 'hash.word');
+        const length = get(context, 'hash.len', 0);
+        const word = get(context, 'hash.word');
 
         if (!Number(length) || length === 0 || length === 1) {
           return word;
         }
 
-        if (_.endsWith(word, 'y')) {
-          return _.replace(word, 'y', 'ies');
+        if (endsWith(word, 'y')) {
+          return replace(word, 'y', 'ies');
         }
 
-        const isMatched = _.compact(
-          _.map(['s', 'x', 'z', 'ch'], (lastChr) => {
-            return _.endsWith(word, lastChr);
+        const isMatched = compact(
+          map(['s', 'x', 'z', 'ch'], (lastChr) => {
+            return endsWith(word, lastChr);
           })
         );
         if (isMatched.length > 0) {
@@ -804,9 +825,9 @@ const Bot = class {
         return word + 's';
       },
       indent: (context) => {
-        const length = _.get(context, 'hash.len', 0);
+        const length = get(context, 'hash.len', 0);
 
-        return _.repeat(' ', length);
+        return repeat(' ', length);
       },
     };
   }
@@ -821,18 +842,18 @@ const Bot = class {
    */
   getSlackUserProfile(identifiers, options = {}) {
     let { email = '0', slackId = '0', userName = '0' } = identifiers;
-    email = _.toUpper(email);
-    const emailIdLocalPart = _.nth(_.split(_.toUpper(email), '@'));
-    const user = _.find(this.getSlackData().members, (member) => {
-      const profileEmail = _.toUpper(_.get(member, 'profile.email'));
+    email = toUpper(email);
+    const emailIdLocalPart = nth(split(toUpper(email), '@'));
+    const user = find(this.getSlackData().members, (member) => {
+      const profileEmail = toUpper(get(member, 'profile.email'));
       return (
         (profileEmail === email ||
-          _.get(member, 'id') === slackId ||
-          _.get(member, 'enterprise_user.id') === slackId ||
-          _.get(member, 'name') === userName ||
+          get(member, 'id') === slackId ||
+          get(member, 'enterprise_user.id') === slackId ||
+          get(member, 'name') === userName ||
           (options.matchLocalPart &&
-            emailIdLocalPart === _.nth(_.split(profileEmail, '@')))) &&
-        !_.get(member, 'deleted')
+            emailIdLocalPart === nth(split(profileEmail, '@')))) &&
+        !get(member, 'deleted')
       );
     });
 
@@ -850,9 +871,9 @@ const Bot = class {
    * @return {string} slack user id.
    */
   getSlackIdFromEmail(emailId) {
-    const uEmailId = _.toUpper(emailId);
-    const user = _.find(this.getSlackData().members, (member) => {
-      return _.toUpper(_.get(member, 'profile.email')) === uEmailId;
+    const uEmailId = toUpper(emailId);
+    const user = find(this.getSlackData().members, (member) => {
+      return toUpper(get(member, 'profile.email')) === uEmailId;
     });
 
     if (user && user.id) {
@@ -869,11 +890,10 @@ const Bot = class {
    * @return {string} slack user id.
    */
   getSlackIdFromEmailLocalPart(emailId) {
-    const uEmailId = _.nth(_.split(_.toUpper(emailId), '@'));
-    const user = _.find(this.getSlackData().members, (member) => {
+    const uEmailId = nth(split(toUpper(emailId), '@'));
+    const user = find(this.getSlackData().members, (member) => {
       return (
-        _.nth(_.split(_.toUpper(_.get(member, 'profile.email', '')), '@')) ===
-        uEmailId
+        nth(split(toUpper(get(member, 'profile.email', '')), '@')) === uEmailId
       );
     });
 
@@ -891,9 +911,9 @@ const Bot = class {
    * @return {string} slack user online presense, away/online.
    */
   getPresenceFromEmail(emailId) {
-    const uEmailId = _.toUpper(emailId);
-    const user = _.find(this.getSlackData().members, (member) => {
-      return _.toUpper(_.get(member, 'profile.email')) === uEmailId;
+    const uEmailId = toUpper(emailId);
+    const user = find(this.getSlackData().members, (member) => {
+      return toUpper(get(member, 'profile.email')) === uEmailId;
     });
 
     if (user && user.presence) {
@@ -910,11 +930,10 @@ const Bot = class {
    * @return {string} slack user online presense, away/online.
    */
   getPresenceFromEmailLocalPart(emailId) {
-    const uEmailId = _.nth(_.split(_.toUpper(emailId), '@'));
-    const user = _.find(this.getSlackData().members, (member) => {
+    const uEmailId = nth(split(toUpper(emailId), '@'));
+    const user = find(this.getSlackData().members, (member) => {
       return (
-        _.nth(_.split(_.toUpper(_.get(member, 'profile.email', '')), '@')) ===
-        uEmailId
+        nth(split(toUpper(get(member, 'profile.email', '')), '@')) === uEmailId
       );
     });
 
@@ -948,7 +967,7 @@ const jsonReplacer = function (key, value) {
  * @param {object} message rename channel message from slack.
  */
 const renameChannel = function (channels, message) {
-  const channel = _.find(channels, {
+  const channel = find(channels, {
     id: message.channel.id,
   });
 
@@ -974,7 +993,7 @@ const channelCreated = function (channels, message) {
  * @param {object} message delete channel message from slack.
  */
 const channelDeleted = function (channels, message) {
-  const deleteChannel = _.remove(channels, (channel) => {
+  const deleteChannel = remove(channels, (channel) => {
     return channel.id == message.channel;
   });
 
@@ -989,7 +1008,7 @@ const channelDeleted = function (channels, message) {
  * @param {object} message user change message from slack.
  */
 const userChange = function (members, message) {
-  let user = _.find(members, {
+  let user = find(members, {
     id: message.user.id,
   });
 
@@ -1005,7 +1024,7 @@ const userChange = function (members, message) {
  * @param {object} message new user message from slack.
  */
 const teamJoin = function (members, message) {
-  members = _.isArray(members) ? members : [];
+  members = isArray(members) ? members : [];
   members.push(message.user);
   logger.debug('Added user ', message.user.id);
 };
@@ -1016,7 +1035,7 @@ const teamJoin = function (members, message) {
  * @param {object} message user presence change message from slack.
  */
 const presenceChange = function (members, message) {
-  const user = _.find(members, {
+  const user = find(members, {
     id: message.user,
   });
 
@@ -1026,6 +1045,4 @@ const presenceChange = function (members, message) {
   }
 };
 
-module.exports = {
-  Bot,
-};
+export { Bot };
