@@ -1,13 +1,25 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
-const handlebars = require('handlebars');
+import * as lodash from 'lodash-es';
+import Handlebars from 'handlebars';
 
-const { Bot } = require('../bot/bot');
-const logger = require('../utils/logger');
+import { Bot } from '../bot/bot.js';
+import logger from '../utils/logger.js';
 
 const ALERT_PARAMS = ['setup', 'list'];
+const {
+  forEach,
+  toUpper,
+  includes,
+  get,
+  isString,
+  map,
+  merge,
+  concat,
+  join,
+  uniq,
+} = lodash;
 
 /**
  *
@@ -34,7 +46,7 @@ const Bots = class {
    * @param {array} bots List of bot in config.
    */
   setup(config, bots) {
-    _.forEach(bots, (bot) => {
+    forEach(bots, (bot) => {
       const newbot = new Bot(normalizeCommand(bot));
 
       if (newbot) {
@@ -70,13 +82,13 @@ const normalizeCommand = function (bot) {
   const stopTasks = [];
   const dataTasks = [];
 
-  _.forEach(bot.botCommand, function (value, key) {
-    const commandKey = _.toUpper(key).replace(/\s/g, '');
+  forEach(bot.botCommand, function (value, key) {
+    const commandKey = toUpper(key).replace(/\s/g, '');
 
     if (value) {
       normalizedCommand[commandKey] = value;
-      _.forEach(value, function (commandAttr, commandAttrkey) {
-        const command = _.toUpper(commandAttr).replace(/\s/g, '');
+      forEach(value, function (commandAttr, commandAttrkey) {
+        const command = toUpper(commandAttr).replace(/\s/g, '');
 
         if (commandAttrkey === 'commandType') {
           value[commandAttrkey] = command;
@@ -84,9 +96,9 @@ const normalizeCommand = function (bot) {
             dataTasks.push(commandKey);
           }
 
-          if (_.includes(['ALERT', 'RECURSIVE'], command)) {
+          if (includes(['ALERT', 'RECURSIVE'], command)) {
             if (command === 'ALERT') {
-              if (_.get(value, 'validation', []).length === 0) {
+              if (get(value, 'validation', []).length === 0) {
                 value.validation = [
                   {
                     schema: [ALERT_PARAMS],
@@ -131,15 +143,15 @@ const normalizeCommand = function (bot) {
           value[commandAttrkey] = command;
         } else if (
           commandAttrkey === 'template' &&
-          _.isString(value[commandAttrkey])
+          isString(value[commandAttrkey])
         ) {
-          value[commandAttrkey] = handlebars.compile(value[commandAttrkey]);
+          value[commandAttrkey] = Handlebars.compile(value[commandAttrkey]);
         } else if (commandAttrkey === 'validation') {
-          value[commandAttrkey] = _.map(value[commandAttrkey], (validation) => {
+          value[commandAttrkey] = map(value[commandAttrkey], (validation) => {
             if (validation.help) {
-              validation.help = _.map(validation.help, (validationHelp) => {
-                return _.merge(validationHelp, {
-                  error: handlebars.compile(validationHelp.error),
+              validation.help = map(validation.help, (validationHelp) => {
+                return merge(validationHelp, {
+                  error: Handlebars.compile(validationHelp.error),
                 });
               });
             }
@@ -155,10 +167,10 @@ const normalizeCommand = function (bot) {
     normalizedCommand['SCHEDULE'] = {
       validation: [
         {
-          schema: _.concat(dataTasks, ['LIST']),
+          schema: concat(dataTasks, ['LIST']),
           help: [
             {
-              recommend: _.concat(dataTasks, ['LIST']),
+              recommend: concat(dataTasks, ['LIST']),
               sample: '{command name} {command args if any} {cron expression}',
               error:
                 'Invalid scehdule command. Use the existing commands ' +
@@ -182,7 +194,7 @@ const normalizeCommand = function (bot) {
     normalizedCommand['STOP'] = normalizedCommand['STOP']
       ? normalizedCommand['STOP']
       : {};
-    _.merge(normalizedCommand['STOP'], {
+    merge(normalizedCommand['STOP'], {
       validation: [
         {
           schema: [stopTasks],
@@ -192,7 +204,7 @@ const normalizeCommand = function (bot) {
               sample: '{command name}',
               error:
                 'Enter a valid command to stop. It should be ' +
-                _.join(stopTasks, ' or '),
+                join(stopTasks, ' or '),
             },
           ],
         },
@@ -220,9 +232,9 @@ const normalizeCommand = function (bot) {
  */
 const mergeAllowedUsers = function (bot) {
   if (bot.allowedUsers) {
-    _.forEach(bot.botCommand, function (command) {
+    forEach(bot.botCommand, function (command) {
       if (!command.allowedUsers) {
-        command.allowedUsers = _.uniq(bot.allowedUsers);
+        command.allowedUsers = uniq(bot.allowedUsers);
       }
     });
   }
@@ -235,14 +247,12 @@ const mergeAllowedUsers = function (bot) {
  */
 const mergeAllowedChannels = function (bot) {
   if (bot.allowedChannels) {
-    _.forEach(bot.botCommand, function (command) {
+    forEach(bot.botCommand, function (command) {
       if (!command.allowedChannels) {
-        command.allowedChannels = _.uniq(bot.allowedChannels);
+        command.allowedChannels = uniq(bot.allowedChannels);
       }
     });
   }
 };
 
-module.exports = {
-  Bots,
-};
+export { Bots };
