@@ -1,12 +1,15 @@
 'use strict';
 
-const _ = require('lodash');
-const http = require('https');
-const querystring = require('querystring');
-const { getUserInfo } = require('../slack-api/users');
+import { request } from 'https';
+import { stringify } from 'querystring';
+import { getUserInfo } from '../slack-api/users.js';
 
-const logger = require('../utils/logger');
-const { Socket } = require('./socket');
+import logger from '../utils/logger.js';
+
+import { Socket } from './socket.js';
+import * as lodash from 'lodash-es';
+
+const { assign, get } = lodash;
 
 const CONFIG = {
   options: {
@@ -67,13 +70,13 @@ const Connector = class {
     let postData = '';
     let options = {};
 
-    postData = querystring.stringify(
-      _.assign(CONFIG.postData, {
+    postData = stringify(
+      assign(CONFIG.postData, {
         token: this.token,
       })
     );
 
-    options = _.assign(CONFIG.options, {
+    options = assign(CONFIG.options, {
       'Content-Length': postData.length,
     });
 
@@ -92,6 +95,7 @@ const Connector = class {
         this.options.socketEventEmitter.emit('connect');
       })
       .catch((err) => {
+        logger.error('Socket connection went bad', err);
         if (err === 'invalid_auth') {
           return Promise.reject(err);
         }
@@ -200,30 +204,27 @@ const Connector = class {
 const handleRequest = function (options, postData) {
   return Promise.resolve({
     then: (onFulfill, onReject) => {
-      const req = http
-        .request(options, function (response) {
-          let responseStr = '';
+      const req = request(options, function (response) {
+        let responseStr = '';
 
-          response.on('data', function (chunk) {
-            responseStr += chunk;
-          });
-
-          response.on('end', function () {
-            const slackData = handleResponse(responseStr);
-
-            if (slackData.error) {
-              return onReject(slackData.error);
-            }
-
-            return onFulfill(slackData);
-          });
-        })
-        .on('error', function (err) {
-          logger.info('Unknown error on connecting to slack ', err);
-          onReject({
-            error: 'unkown_error',
-          });
+        response.on('data', function (chunk) {
+          responseStr += chunk;
         });
+
+        response.on('end', function () {
+          const slackData = handleResponse(responseStr);
+          if (slackData.error) {
+            return onReject(slackData.error);
+          }
+
+          return onFulfill(slackData);
+        });
+      }).on('error', function (err) {
+        logger.info('Unknown error on connecting to slack ', err);
+        onReject({
+          error: 'unkown_error',
+        });
+      });
 
       req.write('' + postData);
       req.end();
@@ -242,25 +243,24 @@ const handleResponse = function (responseStr) {
 
   try {
     slackData = JSON.parse(responseStr);
+    logger.debug('Slack message events', slackData);
   } catch (err) {
     logger.error('Slack response corrupted ', slackData);
     slackData.error = 'invalid_slack_response';
   }
 
-  if (_.get(slackData, 'error') === 'invalid_auth') {
+  if (get(slackData, 'error') === 'invalid_auth') {
     logger.info('Invalid auth token. Unable to connect');
     slackData.error = 'invalid_auth';
-  } else if (_.get(slackData, 'error')) {
-    logger.info('Error connecting to slack ', _.get(slackData, 'error'));
+  } else if (get(slackData, 'error')) {
+    logger.info('Error connecting to slack ', get(slackData, 'error'));
     slackData.error = 'invalid_connection_string';
-  } else if (!_.get(slackData, 'url')) {
-    slackData.error = _.get(slackData, 'error');
+  } else if (!get(slackData, 'url')) {
+    slackData.error = get(slackData, 'error');
     logger.info('Something wrong', slackData.error);
   }
 
   return slackData;
 };
 
-module.exports = {
-  Connector,
-};
+export { Connector };
