@@ -1,7 +1,9 @@
 'use strict';
-const _ = require('lodash');
+import * as lodash from 'lodash-es';
 
-const storage = require('../storage/storage');
+import { Storage } from '../storage/storage.js';
+
+const { get: _get, set: _set, extend } = lodash;
 
 /**
  *
@@ -29,7 +31,7 @@ const EventStore = class {
    */
   get(path, defaultValue) {
     if (path) {
-      return _.get(this.eventStore['runtime'], path, defaultValue);
+      return _get(this.eventStore['runtime'], path, defaultValue);
     }
 
     return this.eventStore['runtime'] || defaultValue;
@@ -43,7 +45,7 @@ const EventStore = class {
    * @return {object} Event data.
    */
   set(path, eventData) {
-    _.set(this.eventStore['runtime'], path, eventData);
+    _set(this.eventStore['runtime'], path, eventData);
 
     return this.get(path);
   }
@@ -54,7 +56,7 @@ const EventStore = class {
    * @return {object} Event data.
    */
   getSchedules() {
-    return _.get(this.eventStore, ['schedule']);
+    return _get(this.eventStore, ['schedule']);
   }
 
   /**
@@ -65,7 +67,7 @@ const EventStore = class {
    * @return {object} Event data.
    */
   setSchedule(path, eventData) {
-    _.set(this.eventStore['schedule'], path, eventData);
+    _set(this.eventStore['schedule'], path, eventData);
     return this.getSchedule();
   }
 
@@ -75,7 +77,7 @@ const EventStore = class {
    * @return {object} Event data.
    */
   getEvents() {
-    return _.get(this.eventStore, ['events']);
+    return _get(this.eventStore, ['events']);
   }
 
   /**
@@ -86,7 +88,7 @@ const EventStore = class {
    * @return {object} Event data.
    */
   setEvents(path, eventData) {
-    _.set(this.eventStore['events'], path, eventData);
+    _set(this.eventStore['events'], path, eventData);
     return this.getEvents();
   }
 
@@ -98,12 +100,13 @@ const EventStore = class {
    * @return {object} Event data.
    */
   update(options, data) {
-    return storage
-      .updateEvents(_.extend({ botName: this.botName }, options), data)
-      .then((eventData) => {
-        this.eventStore[options.eventType] = eventData[this.botName];
-        return eventData[this.botName];
-      });
+    return Storage.updateEvents(
+      extend({ botName: this.botName }, options),
+      data
+    ).then((eventData) => {
+      this.eventStore[options.eventType] = eventData[this.botName];
+      return eventData[this.botName];
+    });
   }
 
   /**
@@ -114,15 +117,14 @@ const EventStore = class {
    * @return {object} Event data.
    */
   remove(options, data) {
-    return storage
-      .removeEvents(_.extend({ botName: this.botName }, options), data)
-      .then((eventData) => {
-        this.eventStore[options.eventType] = eventData[this.botName];
-        return eventData[this.botName];
-      });
+    return Storage.removeEvents(
+      extend({ botName: this.botName }, options),
+      data
+    ).then((eventData) => {
+      this.eventStore[options.eventType] = eventData[this.botName];
+      return eventData[this.botName];
+    });
   }
 };
 
-module.exports = {
-  EventStore,
-};
+export { EventStore };
