@@ -1,9 +1,9 @@
 'use strict';
 
 // Load modules
-const uuid = require('uuid');
-const path = require('path');
-const url = require('url');
+import { v4 } from 'uuid';
+import { join } from 'path';
+import { format } from 'url';
 
 /**
  *
@@ -68,7 +68,7 @@ const Hook = class {
  * @return {string} unique uuid.
  */
 const generateHookId = function () {
-  return uuid.v4();
+  return v4();
 };
 
 /**
@@ -94,11 +94,9 @@ const getHookUrl = function (context, purpose, server) {
 
   urlObj.port = server.address().port;
   urlObj.protocol = server.address().protocol || 'http';
-  urlObj.pathname = path.join('hook', context.botId, purpose.id);
+  urlObj.pathname = join('hook', context.botId, purpose.id);
 
-  return url.format(urlObj, false);
+  return format(urlObj, false);
 };
 
-module.exports = {
-  Hook,
-};
+export { Hook };
