@@ -1,11 +1,11 @@
 'use strict';
-const _ = require('lodash');
+import * as lodash from 'lodash-es';
 
-const { Connector } = require('./connector');
-const { getPort } = require('./socket-server');
+import { Connector } from './connector.js';
+import { getPort } from './socket-server.js';
 
 const localSocketServer = 'ws://0.0.0.0';
-
+const { get, extend } = lodash;
 /**
  *
  * Represents the state and events of a mock connector.
@@ -36,7 +36,7 @@ const MockConnector = class extends Connector {
 
     this.retryAttempt++;
 
-    if (this.retryAttempt >= _.get(this, 'options.mock.retryAttempt', 0)) {
+    if (this.retryAttempt >= get(this, 'options.mock.retryAttempt', 0)) {
       if (this.options.mock.error) {
         return Promise.reject({
           error: this.options.mock.error,
@@ -44,7 +44,7 @@ const MockConnector = class extends Connector {
       }
 
       return Promise.resolve(
-        _.extend(this.options.mock, {
+        extend(this.options.mock, {
           url: localSocketServer + ':' + getPort(),
         })
       );
@@ -56,6 +56,4 @@ const MockConnector = class extends Connector {
   }
 };
 
-module.exports = {
-  MockConnector,
-};
+export { MockConnector };
