@@ -1,23 +1,27 @@
 'use strict';
 
-const _ = require('lodash');
-const fs = require('fs');
-const path = require('path');
-const handlebars = require('handlebars');
+import * as lodash from 'lodash-es';
+import { readFileSync } from 'fs';
+import { join } from 'path';
+import Handlebars from 'handlebars';
 
-const logger = require('../utils/logger');
-const { StreamFile } = require('../bot/stream-file');
-const errorFilePath = path.join(__dirname, '../template/error.hbs');
-const helpFilePath = path.join(__dirname, '../template/help.hbs');
-const alertFilePath = path.join(__dirname, '../template/alert.hbs');
-const errorTemplate = handlebars.compile(
-  fs.readFileSync(errorFilePath, 'utf8')
-);
-const helpTemplate = handlebars.compile(fs.readFileSync(helpFilePath, 'utf8'));
-const alertTemplate = handlebars.compile(
-  fs.readFileSync(alertFilePath, 'utf8')
-);
+import logger from '../utils/logger.js';
+import { StreamFile } from '../bot/stream-file.js';
+import path from 'path';
+import { fileURLToPath } from 'url';
 
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = path.dirname(__filename);
+
+const errorFilePath = join(__dirname, '../template/error.hbs');
+const helpFilePath = join(__dirname, '../template/help.hbs');
+const alertFilePath = join(__dirname, '../template/alert.hbs');
+const errorTemplate = Handlebars.compile(readFileSync(errorFilePath, 'utf8'));
+const helpTemplate = Handlebars.compile(readFileSync(helpFilePath, 'utf8'));
+const alertTemplate = Handlebars.compile(readFileSync(alertFilePath, 'utf8'));
+
+const { isEmpty, filter, set, startsWith, reduce, take, cloneDeep, toUpper } =
+  lodash;
 const CONFIG = {
   commandTypes: ['data', 'recursive', 'kill', 'alert'],
   alertParams: ['setup', 'sample'],
@@ -35,7 +39,7 @@ const generateErrorTemplate = function (botName, botCommand, context) {
   let renderedTemplate = '';
   let commandContext = getContextualErrorResponse(botCommand, context);
 
-  if (_.isEmpty(commandContext)) {
+  if (isEmpty(commandContext)) {
     commandContext = getCommonErrorResponse(botCommand);
   }
 
@@ -49,7 +53,7 @@ const generateErrorTemplate = function (botName, botCommand, context) {
     users: context.users,
     cause: context.cause,
     contextBool:
-      _.filter(commandContext, (o) => {
+      filter(commandContext, (o) => {
         return o.inContext || o.errorContext;
       }).length > 0
         ? true
@@ -75,10 +79,10 @@ const generateErrorTemplate = function (botName, botCommand, context) {
  * @param {object} agent http proxy agent.
  */
 const processFile = function (context, botToken, agent) {
-  _.set(context, 'message.config.botToken', botToken);
-  _.set(context, 'message.config.agent', agent);
-  _.set(context, 'message.config.commandName', context.message.commandName);
-  _.set(context, 'message.config.thread', context.message.thread);
+  set(context, 'message.config.botToken', botToken);
+  set(context, 'message.config.agent', agent);
+  set(context, 'message.config.commandName', context.message.commandName);
+  set(context, 'message.config.thread', context.message.thread);
 
   new StreamFile(context.channels, context.message.data, context.message.config)
     .then((response) => {
@@ -118,7 +122,7 @@ const generateBotResponseTemplate = function (context) {
  * @return {boolean} boolean.
  */
 const isDirectMessage = function (parsedMessage) {
-  return _.startsWith(parsedMessage.channel, 'D', 0);
+  return startsWith(parsedMessage.channel, 'D', 0);
 };
 
 /**
@@ -128,7 +132,7 @@ const isDirectMessage = function (parsedMessage) {
  * @return {boolean} boolean.
  */
 const isPrivateMessage = function (parsedMessage) {
-  return _.startsWith(parsedMessage.channel, 'G', 0);
+  return startsWith(parsedMessage.channel, 'G', 0);
 };
 
 /**
@@ -138,7 +142,7 @@ const isPrivateMessage = function (parsedMessage) {
  * @return {boolean} boolean.
  */
 const isPublicMessage = function (parsedMessage) {
-  return _.startsWith(parsedMessage.channel, 'C', 0);
+  return startsWith(parsedMessage.channel, 'C', 0);
 };
 
 /**
@@ -148,17 +152,17 @@ const isPublicMessage = function (parsedMessage) {
  * @return {object} botCommand with props.
  */
 const getCommonErrorResponse = function (botCommand) {
-  return _.reduce(
+  return reduce(
     botCommand,
     function (item, value, key) {
       item[key] = value;
       item[key].name = key;
 
       if (value.commandType === 'FLOW' && value.validation) {
-        item[key].validation = _.take(value.validation, 2);
+        item[key].validation = take(value.validation, 2);
       }
 
-      if (!_.isEmpty(value.helpText)) {
+      if (!isEmpty(value.helpText)) {
         item[key].helpText = value.helpText;
       }
 
@@ -181,10 +185,10 @@ const getContextualErrorResponse = function (botCommand, context) {
     ? context.parsedMessage.message.params[0]
     : '';
 
-  return _.reduce(
-    _.cloneDeep(
-      _.filter(botCommand, function (n, key) {
-        return _.toUpper(key) === _.toUpper(commandName);
+  return reduce(
+    cloneDeep(
+      filter(botCommand, function (n, key) {
+        return toUpper(key) === toUpper(commandName);
       })
     ),
     function (item, value) {
@@ -194,7 +198,7 @@ const getContextualErrorResponse = function (botCommand, context) {
         item[commandName].commandType === 'FLOW' &&
         item[commandName].validation
       ) {
-        item[commandName].validation = _.take(value.validation, 2);
+        item[commandName].validation = take(value.validation, 2);
       }
 
       item[commandName].name = commandName;
@@ -217,7 +221,7 @@ const getContextualErrorResponse = function (botCommand, context) {
   );
 };
 
-module.exports = {
+export {
   generateErrorTemplate,
   processFile,
   generateAlertResponseTemplate,
