@@ -1,8 +1,8 @@
 'use strict';
 
 // Load modules
-const http = require('http');
-const logger = require('../utils/logger');
+import { createServer } from 'http';
+import logger from '../utils/logger.js';
 
 const DEFAULT_CONFIG = {
   defaultPort: 8080,
@@ -17,7 +17,7 @@ const DEFAULT_CONFIG = {
  * @return {object} Promise object resolve to success or failure.
  */
 const setupServer = function (config, handler) {
-  const server = http.createServer(handler);
+  const server = createServer(handler);
   const port = config.port || DEFAULT_CONFIG.defaultPort;
   const hostname = config.hostname || DEFAULT_CONFIG.defaultHost;
 
@@ -36,6 +36,4 @@ const setupServer = function (config, handler) {
   });
 };
 
-module.exports = {
-  setupServer,
-};
+export { setupServer };
