@@ -1,29 +1,30 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
-const net = require('net');
-const WebSocketServer = require('ws').Server;
+import * as lodash from 'lodash-es';
+import { createServer } from 'net';
+import { WebSocketServer } from 'ws';
 
-const logger = require('../utils/logger');
+import logger from '../utils/logger.js';
 
 const DEFAULT_CONFIG = {
   server: null,
 };
 
+const { get } = lodash;
 /**
  * Function to create mock socket server.
  *
  * @return {object} Promise object resolve to success or failure.
  */
-module.exports.connect = function () {
+export function connect() {
   if (DEFAULT_CONFIG.server !== null) {
     return Promise.resolve(DEFAULT_CONFIG.server);
   }
 
   return Promise.resolve({
     then: (onFulfill) => {
-      const server = net.createServer();
+      const server = createServer();
       server.listen();
 
       server.on('listening', () => {
@@ -46,14 +47,14 @@ module.exports.connect = function () {
       });
     },
   });
-};
+}
 
 /**
  * Function to close all client connections to the socet server.
  *
  * @return {object} Promise object resolve to success or failure.
  */
-module.exports.closeClient = function () {
+export function closeClient() {
   if (!DEFAULT_CONFIG.server) {
     return;
   }
@@ -69,15 +70,15 @@ module.exports.closeClient = function () {
       });
     },
   });
-};
+}
 
-module.exports.getPort = function () {
+export function getPort() {
   if (DEFAULT_CONFIG.server && DEFAULT_CONFIG.server.address()) {
     return DEFAULT_CONFIG.server.address().port;
   }
 
   return;
-};
+}
 
 /**
  * Function to handle incoming messages to the mock server.
@@ -91,16 +92,16 @@ const messageHandler = function (ws, message) {
   try {
     clientMessage = JSON.parse(message);
     /* jshint ignore:start */
-    if (_.get(clientMessage, 'type') === 'message') {
+    if (get(clientMessage, 'type') === 'message') {
       ws.send(
         JSON.stringify({
           ok: true,
-          reply_to: _.get(clientMessage, 'id'),
+          reply_to: get(clientMessage, 'id'),
           ts: Date.now(),
-          message: _.get(clientMessage, 'text', message),
-          type: _.get(clientMessage, 'type', 'message'),
-          user: _.get(clientMessage, 'user', 'U1234567'),
-          channel: _.get(clientMessage, 'channel', 'C1234567'),
+          message: get(clientMessage, 'text', message),
+          type: get(clientMessage, 'type', 'message'),
+          user: get(clientMessage, 'user', 'U1234567'),
+          channel: get(clientMessage, 'channel', 'C1234567'),
         })
       );
     }
