@@ -1,9 +1,10 @@
 'use strict';
 
-const _ = require('lodash');
-const WebSocketPlugin = require('ws');
+import * as lodash from 'lodash-es';
+import WebSocket from 'ws';
+import logger from '../utils/logger.js';
 
-const logger = require('../utils/logger');
+const { extend, get, nth, compact, isEmpty } = lodash;
 
 /**
  *
@@ -38,11 +39,11 @@ const Socket = class {
       then: (onFulfill, onReject) => {
         try {
           if (this.agent) {
-            this.ws = new WebSocketPlugin(this.slackData.url, {
+            this.ws = new WebSocket(this.slackData.url, {
               agent: this.agent,
             });
           } else {
-            this.ws = new WebSocketPlugin(this.slackData.url);
+            this.ws = new WebSocket(this.slackData.url);
           }
 
           return onFulfill();
@@ -62,10 +63,10 @@ const Socket = class {
    * @return {Object} Promise object resolves to Socket instance.
    */
   setupSocketConnectionEvents(slackData) {
-    _.extend(this.slackData, slackData);
-    this.botName = _.get(this.slackData, 'self.name');
-    this.id = _.get(this.slackData, 'self.id');
-    this.eId = _.get(this.slackData, 'self.enterprise_user.id');
+    extend(this.slackData, slackData);
+    this.botName = get(this.slackData, 'self.name');
+    this.id = get(this.slackData, 'self.id');
+    this.eId = get(this.slackData, 'self.enterprise_user.id');
 
     return Promise.all([
       this.connectSocket(),
@@ -94,6 +95,7 @@ const Socket = class {
 
           try {
             slackMessage = JSON.parse(data);
+            logger.debug('Slack socket message events', slackMessage);
           } catch (err) {
             logger.error('Invalid slack data ', data);
           }
@@ -121,15 +123,14 @@ const Socket = class {
     return Promise.resolve({
       then: (onFulfill) => {
         this.ws.on('close', () => {
-          if (this.isShutdown) {
+          logger.debug('Bot connection closed for ', this.botName);
+          if (this.isShutdown || this.id === 'U1234567') {
             this.isShutdown = false;
             this.emitEvent('shutdown');
           } else {
             this.emitEvent('close');
             this.emitEvent('reconnect');
           }
-
-          logger.info('Bot connection closed for ', this.botName);
         });
 
         return onFulfill();
@@ -245,7 +246,7 @@ const Socket = class {
 
       logger.debug(
         'Socket connection closed during ping for ',
-        _.get(this, 'botName')
+        get(this, 'botName')
       );
       this.emitEvent('reconnect');
     }
@@ -272,13 +273,13 @@ const Socket = class {
   routeMessageType(message) {
     const messageTypeRegex = new RegExp(/^(\w+)_\w+$|(\w+)$/);
     const regexResponse = messageTypeRegex.exec(message.type);
-    const messageType = _.nth(_.compact(regexResponse), 1);
+    const messageType = nth(compact(regexResponse), 1);
 
     switch (messageType) {
       case 'message': {
         /* jshint ignore:start */
         const isBotMessage =
-          _.isEmpty(message.reply_to) &&
+          isEmpty(message.reply_to) &&
           !message.subtype &&
           this.ignoreStartupMessage;
         /* jshint ignore:end */
@@ -313,6 +314,4 @@ const Socket = class {
   }
 };
 
-module.exports = {
-  Socket,
-};
+export { Socket };
