@@ -1,12 +1,13 @@
 'use strict';
 
-const _ = require('lodash');
-const plot = require('@umashankar/plotter').plot;
-const logger = require('../utils/logger');
-const postAttachmentFromStream = require('../slack-api/post-attachment')
-  .postAttachmentFromStream;
+import * as lodash from 'lodash-es';
+import { plot } from '@usubram/plotter';
+import logger from '../utils/logger.js';
+import { postAttachmentFromStream } from '../slack-api/post-attachment.js';
 
 const graphFileTypes = ['png', 'svg'];
+
+const { merge, includes } = lodash;
 
 /**
  *
@@ -23,9 +24,9 @@ const StreamFile = class {
    */
   constructor(channel, data, config) {
     if (data.responseType) {
-      config = _.merge(config, data.responseType);
+      config = merge(config, data.responseType);
     }
-    if (_.includes(graphFileTypes, config.type) && config.style) {
+    if (includes(graphFileTypes, config.type) && config.style) {
       return handleGraphResponse(channel, data, config);
     } else {
       return handleFileResponse(channel, data, config);
@@ -114,6 +115,4 @@ const generateGraph = function (plotConfig) {
   }
 };
 
-module.exports = {
-  StreamFile,
-};
+export { StreamFile };
