@@ -1,16 +1,17 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
+import * as lodash from 'lodash-es';
+import logger from '../utils/logger.js';
+import { Data } from '../command/commands/data.js';
+import { Recursive } from '../command/commands/recursive.js';
+import { Alert } from '../command/commands/alert.js';
+import { Flow } from '../command/commands/flow.js';
+import { Kill } from '../command/commands/kill.js';
+import { Schedule } from '../command/commands/schedule.js';
 
-const logger = require('../utils/logger');
-const { Data } = require('../command/commands/data');
-const { Recursive } = require('../command/commands/recursive');
-const { Alert } = require('../command/commands/alert');
-const { Flow } = require('../command/commands/flow');
-const { Kill } = require('../command/commands/kill');
-const { Schedule } = require('../command/commands/schedule');
-
+const { reduce, concat, values, toUpper, get, head, compact, map, isFunction } =
+  lodash;
 /**
  *
  * Represents the state and events of command factory.
@@ -41,7 +42,7 @@ const CommandFactory = class {
    * @return {object} Resolves to this instance.
    */
   loadCommands() {
-    return _.reduce(
+    return reduce(
       this.options.getBotConfig().botCommand,
       (result, command, key) => {
         return this.getCommand(
@@ -105,16 +106,17 @@ const CommandFactory = class {
    * @return {object} Resolves to loaded commands.
    */
   loadEvents() {
-    const savedEvents = _.concat(
-      _.values(this.options.getEventStore().getSchedules()),
-      _.values(this.options.getEventStore().getEvents())
+    const savedEvents = concat(
+      values(this.options.getEventStore().getSchedules()),
+      values(this.options.getEventStore().getEvents())
     );
 
     if (savedEvents) {
       return savedEvents.reduce((promiseItem, savedEvent) => {
-        const command = this.commandObj[
-          _.toUpper(_.get(savedEvent, 'parsedMessage.message.command'))
-        ];
+        const command =
+          this.commandObj[
+            toUpper(get(savedEvent, 'parsedMessage.message.command'))
+          ];
         if (command) {
           return command.reloadCommand(savedEvent.parsedMessage);
         }
@@ -133,9 +135,7 @@ const CommandFactory = class {
   handleMessage(parsedMessage) {
     return Promise.resolve({
       then: (onFulfill, onReject) => {
-        const command = this.commandObj[
-          _.get(parsedMessage, 'message.command')
-        ];
+        const command = this.commandObj[get(parsedMessage, 'message.command')];
         if (command) {
           return command
             .validate(parsedMessage)
@@ -169,22 +169,22 @@ const CommandFactory = class {
   handleHook(purposeId, requestData) {
     return Promise.resolve({
       then: (onFulfill, onReject) => {
-        const hookInstance = _.head(
-          _.compact(
-            _.map(
+        const hookInstance = head(
+          compact(
+            map(
               this.commandObj,
               ['hookContext', purposeId, 'command'].join('.')
             )
           )
         );
-        const commandModel = _.get(this.commandObj, hookInstance, undefined);
-        const template = _.get(commandModel, 'template');
+        const commandModel = get(this.commandObj, hookInstance, undefined);
+        const template = get(commandModel, 'template');
         let renderedData = requestData ? requestData.text || requestData : '';
 
         if (requestData && hookInstance && commandModel) {
           try {
             renderedData =
-              !requestData.text && _.isFunction(template)
+              !requestData.text && isFunction(template)
                 ? template(requestData)
                 : renderedData;
           } catch (err) {
@@ -204,6 +204,4 @@ const CommandFactory = class {
   }
 };
 
-module.exports = {
-  CommandFactory,
-};
+export { CommandFactory };
