@@ -1,11 +1,48 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
+import * as lodash from 'lodash-es';
 
-const logger = require('../utils/logger');
-const responseHandler = require('../bot/response-handler');
+import logger from '../utils/logger.js';
+import {
+  isPublicMessage,
+  isPrivateMessage,
+  isDirectMessage,
+} from '../bot/response-handler.js';
 
+const {
+  find,
+  get,
+  compact,
+  zipWith,
+  toUpper,
+  unset,
+  set,
+  trim,
+  includes,
+  filter,
+  values,
+  map,
+  join,
+  isEmpty,
+  cloneDeep,
+  orderBy,
+  merge,
+  assign,
+  isArray,
+  isRegExp,
+  last,
+  first,
+  pick,
+  isFunction,
+  nth,
+  some,
+  indexOf,
+  toString,
+  flatten,
+  isUndefined,
+  isNull,
+} = lodash;
 const SCHEDULE_COMMAND = {
   LIST: 'LIST',
 };
@@ -228,13 +265,13 @@ const Command = class {
    * @return {object}
    */
   buildOptions(slackResponse, slackData, purpose) {
-    const user = _.find(slackData.members, {
+    const user = find(slackData.members, {
       id: slackResponse.user,
     });
 
     return {
       channel: slackResponse.channel,
-      hookUrl: _.get(purpose, 'url'),
+      hookUrl: get(purpose, 'url'),
       user: user || { id: slackResponse.user },
     };
   }
@@ -247,10 +284,10 @@ const Command = class {
    * @return {object} Assigned default values.
    */
   setDefaultParams(defaultValues, slackResponse) {
-    const param = _.compact(getParams(slackResponse, null, true));
+    const param = compact(getParams(slackResponse, null, true));
 
     if (defaultValues) {
-      const assignedDefault = _.zipWith(param, defaultValues, (a, b) => {
+      const assignedDefault = zipWith(param, defaultValues, (a, b) => {
         return a || b;
       });
 
@@ -308,7 +345,7 @@ const Command = class {
    * @return {string} command name.
    */
   getCommandName(commandName) {
-    return _.toUpper(commandName || this.commandName);
+    return toUpper(commandName || this.commandName);
   }
 
   /**
@@ -317,8 +354,8 @@ const Command = class {
    * @return {object} handlebar template.
    */
   getTemplate() {
-    let template = this.getBotConfig().botCommand[this.getCommandName()]
-      .template;
+    let template =
+      this.getBotConfig().botCommand[this.getCommandName()].template;
 
     try {
       template = template || '';
@@ -341,11 +378,11 @@ const Command = class {
 
     if (timer) {
       clearInterval(timer);
-      _.unset(this.getEventStore().get(), path);
+      unset(this.getEventStore().get(), path);
     }
 
-    if (!_.get(this.getEventStore().get(), path)) {
-      _.set(
+    if (!get(this.getEventStore().get(), path)) {
+      set(
         this.getEventStore().get(),
         path,
         setInterval(() => {
@@ -364,7 +401,7 @@ const Command = class {
    * @return {object} Timer instance.
    */
   getTimer(parsedMessage, path) {
-    return _.get(this.getEventStore().get(), path);
+    return get(this.getEventStore().get(), path);
   }
 
   /**
@@ -377,25 +414,22 @@ const Command = class {
     const currentCommand = this.getCommand(
       this.getCommandName(this.commandName)
     );
-    const channel = _.get(parsedMessage, 'channel');
-    const options = _.trim(this.getParams(parsedMessage, 1));
-    const isPublicRequest = responseHandler.isPublicMessage(parsedMessage);
-    const isPrivateRequest = responseHandler.isPrivateMessage(parsedMessage);
-    const isDirectRequest = responseHandler.isDirectMessage(parsedMessage);
-    const events = _.includes(
-      ['ALERT', 'RECURSIVE'],
-      currentCommand.commandType
-    )
+    const channel = get(parsedMessage, 'channel');
+    const options = trim(this.getParams(parsedMessage, 1));
+    const isPublicRequest = isPublicMessage(parsedMessage);
+    const isPrivateRequest = isPrivateMessage(parsedMessage);
+    const isDirectRequest = isDirectMessage(parsedMessage);
+    const events = includes(['ALERT', 'RECURSIVE'], currentCommand.commandType)
       ? this.getEventStore().getEvents()
       : this.getEventStore().getSchedules();
 
-    const allowedEvents = _.filter(_.values(events), (item) => {
-      const isSameChannel = _.get(item, 'parsedMessage.channel') === channel;
-      const isSchedule = !!_.get(item, 'parsedMessage.scheduleId');
+    const allowedEvents = filter(values(events), (item) => {
+      const isSameChannel = get(item, 'parsedMessage.channel') === channel;
+      const isSchedule = !!get(item, 'parsedMessage.scheduleId');
       const isSameUser =
-        _.get(item, 'parsedMessage.user') === _.get(parsedMessage, 'user');
+        get(item, 'parsedMessage.user') === get(parsedMessage, 'user');
       const isPublicSchedule = isSchedule
-        ? responseHandler.isPublicMessage(item.parsedMessage)
+        ? isPublicMessage(item.parsedMessage)
         : false;
 
       if ((isPublicRequest || isPrivateRequest) && options === 'ALL') {
@@ -409,13 +443,13 @@ const Command = class {
       }
     });
 
-    return _.map(allowedEvents, (item) => {
+    return map(allowedEvents, (item) => {
       const context = {
-        user: _.get(item, 'parsedMessage.user'),
-        command: _.join(_.get(item, 'parsedMessage.message.params'), ' '),
-        channel: _.get(item, 'parsedMessage.channel'),
-        private: !responseHandler.isPublicMessage(_.get(item, 'parsedMessage')),
-        scheduleId: _.get(item, 'parsedMessage.scheduleId'),
+        user: get(item, 'parsedMessage.user'),
+        command: join(get(item, 'parsedMessage.message.params'), ' '),
+        channel: get(item, 'parsedMessage.channel'),
+        private: !isPublicMessage(get(item, 'parsedMessage')),
+        scheduleId: get(item, 'parsedMessage.scheduleId'),
         botName: this.getEventStore().botName,
       };
 
@@ -448,7 +482,7 @@ const Command = class {
       failedParams: {},
     };
 
-    if (_.isEmpty(command.validation)) {
+    if (isEmpty(command.validation)) {
       result.isNoop = true;
       return result;
     }
@@ -458,14 +492,15 @@ const Command = class {
       return result;
     }
 
-    const commandValidations = _.cloneDeep(
+    const commandValidations = cloneDeep(
       command.commandType === 'FLOW'
         ? this.getFlowValidations(parsedMessage)
-        : command.validation);
+        : command.validation
+    );
 
-    const validationResult = _.compact(
-      _.orderBy(
-        _.map(commandValidations, (rule) => {
+    const validationResult = compact(
+      orderBy(
+        map(commandValidations, (rule) => {
           let validationResultValue = 0;
           let noOfErrors = 0;
 
@@ -474,21 +509,21 @@ const Command = class {
             parsedMessage
           );
 
-          return _.merge(
+          return merge(
             {},
             rule,
             {
-              help: _.map(
+              help: map(
                 rule.schema,
                 (schema, index) => {
                   const inputParam = defaultParams[index] || ' ';
-                  const help = _.assign({}, rule.help[index]);
+                  const help = assign({}, rule.help[index]);
                   const position = index + 1;
 
-                  if (_.isArray(schema)) {
+                  if (isArray(schema)) {
                     result.isMultiParam = true;
                     help.result = this.validateArgument(schema, inputParam);
-                  } else if (_.isRegExp(schema)) {
+                  } else if (isRegExp(schema)) {
                     result.isMultiParam = true;
                     help.result = this.validateArgument(schema, inputParam);
                   } else {
@@ -516,7 +551,7 @@ const Command = class {
       )
     );
 
-    const validResult = _.filter(
+    const validResult = filter(
       validationResult,
       (result) => result.noOfErrors === 0
     );
@@ -526,21 +561,21 @@ const Command = class {
 
     const targetError =
       command.commandType === 'FLOW'
-        ? _.last(validationResult)
-        : _.first(validationResult);
+        ? last(validationResult)
+        : first(validationResult);
 
     if (!result.isValid) {
-      _.merge(result, {
-        noOfErrors: _.get(targetError, 'noOfErrors'),
-        cause: _.pick(targetError, ['schema', 'help']),
-        failedParams: _.compact(
-          _.map(targetError.help, (item) => {
-            if (item.result.failed && !_.isEmpty(_.trim(item.result.input))) {
+      merge(result, {
+        noOfErrors: get(targetError, 'noOfErrors'),
+        cause: pick(targetError, ['schema', 'help']),
+        failedParams: compact(
+          map(targetError.help, (item) => {
+            if (item.result.failed && !isEmpty(trim(item.result.input))) {
               const error = item.error({
                 arg: item.result.input,
               });
               return {
-                error: _.isFunction(item.error) ? error : item.error,
+                error: isFunction(item.error) ? error : item.error,
               };
             } else if (item.result.failed) {
               return {
@@ -552,7 +587,7 @@ const Command = class {
       });
     } else {
       parsedMessage.message.params = this.setDefaultParams(
-        _.nth(validResult || []).default,
+        nth(validResult || []).default,
         parsedMessage
       );
     }
@@ -573,27 +608,24 @@ const Command = class {
       failed: true,
     };
 
-    if (_.isRegExp(schema) && schema.test(input)) {
+    if (isRegExp(schema) && schema.test(input)) {
       result.failed = false;
       return result;
     }
 
     if (
-      _.isArray(schema) &&
-      _.some(schema, (item) => {
-        return _.isRegExp(item) && item.test(input);
+      isArray(schema) &&
+      some(schema, (item) => {
+        return isRegExp(item) && item.test(input);
       })
     ) {
       result.failed = false;
       return result;
     }
 
-    const op1 = _.indexOf(
-      _.map(schema, _.toUpper),
-      _.toString(_.toUpper(input))
-    );
-    const op2 = _.indexOf(schema, input);
-    const op3 = _.toUpper(schema) === _.toUpper(input);
+    const op1 = indexOf(map(schema, toUpper), toString(toUpper(input)));
+    const op2 = indexOf(schema, input);
+    const op3 = toUpper(schema) === toUpper(input);
 
     if (op1 > -1 || op2 > -1 || op3) {
       result.failed = false;
@@ -614,7 +646,7 @@ const Command = class {
   isCommandAllowed(command, slackResponse, slackData) {
     if (command && command.allowedUsers) {
       return (
-        _.includes(command.allowedUsers, slackResponse.user) ||
+        includes(command.allowedUsers, slackResponse.user) ||
         this.validateGroup(
           slackResponse.user,
           command.allowedUsers,
@@ -635,7 +667,7 @@ const Command = class {
    */
   isAllowedChannel(command, slackResponse) {
     if (command && command.allowedChannels) {
-      return _.includes(command.allowedChannels, slackResponse.channel);
+      return includes(command.allowedChannels, slackResponse.channel);
     }
 
     return true;
@@ -652,7 +684,7 @@ const Command = class {
     if (
       command &&
       command.blockDirectMessage &&
-      responseHandler.isDirectMessage(slackResponse)
+      isDirectMessage(slackResponse)
     ) {
       return command.blockDirectMessage;
     }
@@ -670,15 +702,15 @@ const Command = class {
    */
   validateGroup(currentUser, allowedUsers, usergroups) {
     if (usergroups) {
-      const filteredUserGroup = _.flatten(
-        _.map(allowedUsers, function (allowedUser) {
-          return _.find(usergroups, {
+      const filteredUserGroup = flatten(
+        map(allowedUsers, function (allowedUser) {
+          return find(usergroups, {
             handle: allowedUser,
           });
         })
       );
 
-      return _.includes(filteredUserGroup, currentUser);
+      return includes(filteredUserGroup, currentUser);
     }
 
     return false;
@@ -686,8 +718,8 @@ const Command = class {
 };
 
 const getParams = function (slackResponse, level, realValue = false) {
-  const paramsLen = _.get(slackResponse, 'message.params', []).length;
-  if (!paramsLen && (_.isUndefined(level) || _.isNull(level))) {
+  const paramsLen = get(slackResponse, 'message.params', []).length;
+  if (!paramsLen && (isUndefined(level) || isNull(level))) {
     return [];
   }
 
@@ -696,22 +728,20 @@ const getParams = function (slackResponse, level, realValue = false) {
       level = paramsLen === 1 ? 0 : paramsLen - 1;
     }
 
-    if (_.isUndefined(level) || _.isNull(level)) {
+    if (isUndefined(level) || isNull(level)) {
       if (realValue) {
-        return _.get(slackResponse, 'message.params', []);
+        return get(slackResponse, 'message.params', []);
       }
 
-      return _.map(_.get(slackResponse, 'message.params', [], _.toUpper));
+      return map(get(slackResponse, 'message.params', [], toUpper));
     }
 
     if (realValue) {
-      return _.get(slackResponse, 'message.params[' + level + ']');
+      return get(slackResponse, 'message.params[' + level + ']');
     }
 
-    return _.toUpper(_.get(slackResponse, 'message.params[' + level + ']'));
+    return toUpper(get(slackResponse, 'message.params[' + level + ']'));
   }
 };
 
-module.exports = {
-  Command,
-};
+export { Command };
