@@ -1,14 +1,40 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
-const moment = require('moment');
-const shortid = require('shortid');
+import * as lodash from 'lodash-es';
+import * as moment from 'moment';
+import { nanoid } from 'nanoid';
 
-const logger = require('../../utils/logger');
-const { Command } = require('../command');
-const responseHandler = require('../../bot/response-handler');
+import logger from '../../utils/logger.js';
+import { Command } from '../command.js';
+import {
+  generateBotResponseTemplate,
+  generateAlertResponseTemplate,
+} from '../../bot/response-handler.js';
 
+const { unix } = moment;
+const {
+  toLower,
+  get,
+  filter,
+  values,
+  isEqual,
+  isNumber,
+  each,
+  partial,
+  isFunction,
+  set,
+  merge,
+  concat,
+  endsWith,
+  extend,
+  sortBy,
+  last,
+  first,
+  size,
+  isArray,
+  slice,
+} = lodash;
 const DEFAULT_CONFIG = {
   alertParams: ['SETUP', 'LIST'],
   algo: {
@@ -58,20 +84,20 @@ const Alert = class extends Command {
           const currentParams = (this.getParams(parsedMessage) || []).join('');
 
           parsedMessage.scheduleId =
-            parsedMessage.scheduleId || _.toLower(shortid.generate());
+            parsedMessage.scheduleId || toLower(nanoid(6));
 
           this.alertTaskChannelPath = [this.commandName, 'tasks'];
           this.alertTaskTimerPath = [this.commandName, 'timer'];
 
-          const alertEvents = _.get(
+          const alertEvents = get(
             this.getEventStore().get(),
             this.alertTaskChannelPath
           );
 
-          const isAlertExist = _.filter(_.values(alertEvents), (alertEvent) => {
+          const isAlertExist = filter(values(alertEvents), (alertEvent) => {
             return (
               alertEvent &&
-              _.isEqual(
+              isEqual(
                 currentParams,
                 (this.getParams(alertEvent.parsedMessage) || []).join('')
               ) &&
@@ -82,7 +108,7 @@ const Alert = class extends Command {
           if (isAlertExist.length > 0) {
             this.messageHandler({
               channels: parsedMessage.channel,
-              message: responseHandler.generateBotResponseTemplate({
+              message: generateBotResponseTemplate({
                 parsedMessage: parsedMessage,
                 alertExist: true,
                 commandName: this.commandName,
@@ -91,11 +117,13 @@ const Alert = class extends Command {
               thread: parsedMessage.thread_ts,
             });
 
-            return onReject(new Error('Alert with arguments already exist'));
+            return onReject(
+              new logger.error('Alert with arguments already exist')
+            );
           }
 
           let time = this.getParams(parsedMessage, 0);
-          time = _.isNumber(time) && time > 0 ? time : 1;
+          time = isNumber(time) && time > 0 ? time : 1;
           if (this.getCommand().timeUnit === 'h') {
             time = time * 3600000;
           } else {
@@ -155,7 +183,7 @@ const Alert = class extends Command {
 
           this.messageHandler({
             channels: parsedMessage.channel,
-            message: responseHandler.generateBotResponseTemplate({
+            message: generateBotResponseTemplate({
               parsedMessage: parsedMessage,
               noAlertEvents: noEvents,
               alertEvents: scheduledEvents,
@@ -168,7 +196,7 @@ const Alert = class extends Command {
           return onFulfill();
         }
 
-        const alertEvents = _.get(
+        const alertEvents = get(
           this.getEventStore().get(),
           this.alertTaskChannelPath
         );
@@ -177,7 +205,7 @@ const Alert = class extends Command {
           this.message.apply(this, [err, data, eventKey]);
         };
 
-        _.each(_.values(alertEvents), (alertEvent) => {
+        each(values(alertEvents), (alertEvent) => {
           if (typeof this.getCommand().snooze === 'function') {
             const shouldSnooze = this.getCommand().snooze.apply(this, [
               {
@@ -187,7 +215,7 @@ const Alert = class extends Command {
             ]);
 
             if (shouldSnooze) {
-              logger.log('Snoozing alert', { time: new Date() });
+              logger.info('Snoozing alert', { time: new Date() });
               return;
             }
           }
@@ -202,7 +230,7 @@ const Alert = class extends Command {
               this.getSlackData(),
               this.purpose
             ),
-            _.partial(callback, alertEvent.parsedMessage.scheduleId),
+            partial(callback, alertEvent.parsedMessage.scheduleId),
           ]);
         });
 
@@ -262,8 +290,8 @@ const Alert = class extends Command {
         ) {
           this.messageHandler({
             channels: parsedMessage.channel,
-            message: responseHandler.generateBotResponseTemplate({
-              template: _.isFunction(this.template) ? this.template : '',
+            message: generateBotResponseTemplate({
+              template: isFunction(this.template) ? this.template : '',
               /* jshint ignore:start */
               alertNotification: true,
               /* jshint ignore:end */
@@ -317,15 +345,15 @@ const Alert = class extends Command {
     let events = {};
     const channels = {};
 
-    _.set(channels, parsedMessage.scheduleId, {
+    set(channels, parsedMessage.scheduleId, {
       parsedMessage: parsedMessage,
       sensitivity: sensitivity,
       commandType: this.getCommand().commandType,
       handler: this.getCommand().data,
     });
-    events = _.merge(
+    events = merge(
       {},
-      _.get(this.getEventStore().get(), this.alertTaskChannelPath, {}),
+      get(this.getEventStore().get(), this.alertTaskChannelPath, {}),
       channels
     );
 
@@ -354,25 +382,25 @@ const Alert = class extends Command {
    */
   handleCumulativeDifference(data, eventKey) {
     if (data) {
-      const alertData = _.get(
+      const alertData = get(
         this.getEventStore().get(),
-        _.concat(this.alertTaskChannelPath, eventKey)
+        concat(this.alertTaskChannelPath, eventKey)
       );
-      const sensitivity = _.get(alertData, 'sensitivity');
-      const isPercentage = _.endsWith(sensitivity, '%');
+      const sensitivity = get(alertData, 'sensitivity');
+      const isPercentage = endsWith(sensitivity, '%');
       const cResult = computeCumulativeDifference(
         data,
-        _.get(alertData, 'previousDataSet'),
+        get(alertData, 'previousDataSet'),
         isPercentage
       );
 
       alertData.previousDataSet = cResult.previousDataSet;
       this.getEventStore().set(
-        _.concat(this.alertTaskChannelPath, eventKey),
+        concat(this.alertTaskChannelPath, eventKey),
         alertData
       );
 
-      const result = _.extend(
+      const result = extend(
         {
           template: this.template,
           cumulativeDifferenceAlert: true,
@@ -386,7 +414,7 @@ const Alert = class extends Command {
       if (Math.abs(result.difference) > parseInt(sensitivity, 10)) {
         this.messageHandler({
           channels: [alertData.parsedMessage.channel],
-          message: responseHandler.generateAlertResponseTemplate(result),
+          message: generateAlertResponseTemplate(result),
         });
       }
     }
@@ -403,14 +431,14 @@ const Alert = class extends Command {
       const varianceResult = calculateVariance(data);
 
       if (varianceResult && varianceResult.perct) {
-        alertData = _.get(this, _.concat(this.alertTaskChannelPath, eventKey));
+        alertData = get(this, concat(this.alertTaskChannelPath, eventKey));
       }
 
-      if (varianceResult.perct > _.get(alertData, 'sensitivity')) {
+      if (varianceResult.perct > get(alertData, 'sensitivity')) {
         this.messageHandler({
           channels: [alertData.parsedMessage.channel],
-          message: responseHandler.generateAlertResponseTemplate(
-            _.extend(
+          message: generateAlertResponseTemplate(
+            extend(
               {
                 template: this.template,
                 consistentVariationAlert: true,
@@ -482,18 +510,17 @@ const computeCumulativeDifference = function (
   previousDataSet,
   isPercentage
 ) {
-  const sortedDataSetByTime = _.sortBy(timeSeriesDataSet, 'time');
-  const lastPreviousData =
-    _.last(previousDataSet) || _.last(sortedDataSetByTime);
-  const currentDataTime = _.get(_.last(sortedDataSetByTime), 'time');
+  const sortedDataSetByTime = sortBy(timeSeriesDataSet, 'time');
+  const lastPreviousData = last(previousDataSet) || last(sortedDataSetByTime);
+  const currentDataTime = get(last(sortedDataSetByTime), 'time');
   const isFuture = currentDataTime
-    ? moment.unix(currentDataTime).isAfter(moment.unix(lastPreviousData.time))
+    ? unix(currentDataTime).isAfter(unix(lastPreviousData.time))
     : false;
   let result = {};
 
   if (lastPreviousData && isFuture) {
-    const lastValue = _.last(sortedDataSetByTime).value;
-    const firstValue = _.first(sortedDataSetByTime).value;
+    const lastValue = last(sortedDataSetByTime).value;
+    const firstValue = first(sortedDataSetByTime).value;
     const difference = lastValue - firstValue;
     const isHigher = lastValue > firstValue;
 
@@ -522,7 +549,7 @@ const computeCumulativeDifference = function (
   }
 
   result.previousDataSet =
-    _.size(sortedDataSetByTime) > 0 ? sortedDataSetByTime : previousDataSet;
+    size(sortedDataSetByTime) > 0 ? sortedDataSetByTime : previousDataSet;
 
   return result;
 };
@@ -534,15 +561,15 @@ const computeCumulativeDifference = function (
  * @return {object} computed percentage difference.
  */
 const calculateVariance = function (dataSetArr) {
-  const dataSetLength = _.isArray(dataSetArr) ? dataSetArr.length : 0;
+  const dataSetLength = isArray(dataSetArr) ? dataSetArr.length : 0;
   let sensitivityPercentage = 0;
 
   if (dataSetLength > 0 && dataSetLength % 2 === 0) {
     const sdSet1 = Math.abs(
-      standardDeviation(_.slice(dataSetArr, 0, dataSetLength / 2))
+      standardDeviation(slice(dataSetArr, 0, dataSetLength / 2))
     );
     const sdSet2 = Math.abs(
-      standardDeviation(_.slice(dataSetArr, dataSetLength / 2))
+      standardDeviation(slice(dataSetArr, dataSetLength / 2))
     );
 
     if (sdSet1 !== 0 && sdSet2 !== 0) {
@@ -596,6 +623,4 @@ const average = function (values) {
   return sum / values.length;
 };
 
-module.exports = {
-  Alert,
-};
+export { Alert };
