@@ -1,13 +1,14 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
+import * as lodash from 'lodash-es';
 
-const logger = require('../../utils/logger');
-const { Command } = require('../command');
-const responseHandler = require('../../bot/response-handler');
-const postMessage = require('../../slack-api/post-message');
+import logger from '../../utils/logger.js';
+import { Command } from '../command.js';
+import { processFile } from '../../bot/response-handler.js';
+import postMessage from '../../slack-api/post-message.js';
 
+const { isEmpty, get, isFunction } = lodash;
 /**
  *
  * Represents the state and events of a data command.
@@ -44,7 +45,7 @@ const Data = class extends Command {
     return Promise.resolve({
       then: (onFulfill, onReject) => {
         this.template = '';
-        this.purpose = _.isEmpty(this.getHook())
+        this.purpose = isEmpty(this.getHook())
           ? ''
           : this.getHook().getHookPurpose(parsedMessage.channel);
         this.hookContext = this.getHookContext(
@@ -136,8 +137,8 @@ const Data = class extends Command {
       return;
     }
 
-    if (this.getCommand().responseType || _.get(data, 'responseType')) {
-      if (_.get(data, 'responseType') === 'rich') {
+    if (this.getCommand().responseType || get(data, 'responseType')) {
+      if (get(data, 'responseType') === 'rich') {
         postMessage(
           {
             botToken: this.getBotConfig().botToken,
@@ -146,13 +147,13 @@ const Data = class extends Command {
           Object.assign(data.response, {
             channel: parsedMessage.channel,
             thread_ts:
-              parsedMessage.thread_ts || _.get(data, 'thread')
+              parsedMessage.thread_ts || get(data, 'thread')
                 ? parsedMessage.thread_ts || parsedMessage.ts
                 : '',
           })
         );
       } else {
-        responseHandler.processFile(
+        processFile(
           {
             channels: [parsedMessage.channel],
             message: {
@@ -166,7 +167,7 @@ const Data = class extends Command {
           this.getHttpAgent()
         );
       }
-    } else if (data && _.isFunction(this.template)) {
+    } else if (data && isFunction(this.template)) {
       try {
         const renderedData = this.template(data);
 
@@ -198,6 +199,4 @@ const Data = class extends Command {
   }
 };
 
-module.exports = {
-  Data,
-};
+export { Data };
