@@ -1,12 +1,16 @@
 'use strict';
 
-const _ = require('lodash');
+import * as lodash from 'lodash-es';
 
-const logger = require('../../utils/logger');
-const { Command } = require('../command');
-const responseHandler = require('../../bot/response-handler');
-const postMessage = require('../../slack-api/post-message');
+import logger from '../../utils/logger.js';
+import { Command } from '../command.js';
+import {
+  generateBotResponseTemplate,
+  processFile,
+} from '../../bot/response-handler.js';
+import postMessage from '../../slack-api/post-message.js';
 
+const { isEmpty, get, compact, isFunction, find, take } = lodash;
 /**
  *
  * Represents the state and events of a flow command.
@@ -43,7 +47,7 @@ const Flow = class extends Command {
     return Promise.resolve({
       then: (onFulfill, onReject) => {
         this.template = '';
-        this.purpose = _.isEmpty(this.getHook())
+        this.purpose = isEmpty(this.getHook())
           ? ''
           : this.getHook().getHookPurpose(parsedMessage.channel);
         this.hookContext = this.getHookContext(
@@ -61,7 +65,7 @@ const Flow = class extends Command {
           'timer',
         ];
 
-        const timer = _.get(this.getEventStore().get(), this.timerKey);
+        const timer = get(this.getEventStore().get(), this.timerKey);
 
         this.timeout =
           Number(this.getCommand().timeout) > 0
@@ -76,7 +80,7 @@ const Flow = class extends Command {
         ) {
           question = this.getEventStore().set(
             this.questionKey,
-            _.compact(
+            compact(
               this.getCommand().validation.map((item) =>
                 Object.assign({}, item)
               )
@@ -128,7 +132,7 @@ const Flow = class extends Command {
 
           this.messageHandler({
             channels: parsedMessage.channel,
-            message: responseHandler.generateBotResponseTemplate({
+            message: generateBotResponseTemplate({
               parsedMessage: parsedMessage,
               noflow: flow,
               flowExist: !flow,
@@ -145,7 +149,7 @@ const Flow = class extends Command {
         if (!timer) {
           this.messageHandler({
             channels: parsedMessage.channel,
-            message: responseHandler.generateBotResponseTemplate({
+            message: generateBotResponseTemplate({
               parsedMessage: parsedMessage,
               flowExpired: true,
               botName: this.getEventStore().botName,
@@ -218,12 +222,12 @@ const Flow = class extends Command {
     }
 
     let renderedData = {};
-    if (_.get(data, 'terminated') === true) {
+    if (get(data, 'terminated') === true) {
       const flow = this.isFlowExist();
       this.invalidateFlow(this.questionKey, this.timerKey);
 
-      if (_.get(data, 'error')) {
-        if (_.isFunction(this.template)) {
+      if (get(data, 'error')) {
+        if (isFunction(this.template)) {
           data._response = {
             response: data.response,
           };
@@ -234,7 +238,7 @@ const Flow = class extends Command {
             message: renderedData,
             thread: parsedMessage.thread_ts,
           });
-        } else if (_.get(data, 'responseType') === 'rich') {
+        } else if (get(data, 'responseType') === 'rich') {
           postMessage(
             {
               botToken: this.getBotConfig().botToken,
@@ -243,7 +247,7 @@ const Flow = class extends Command {
             Object.assign(data.response, {
               channel: parsedMessage.channel,
               thread_ts:
-                parsedMessage.thread_ts || _.get(data, 'thread')
+                parsedMessage.thread_ts || get(data, 'thread')
                   ? parsedMessage.thread_ts || parsedMessage.ts
                   : '',
             })
@@ -252,7 +256,7 @@ const Flow = class extends Command {
       } else {
         this.messageHandler({
           channels: parsedMessage.channel,
-          message: responseHandler.generateBotResponseTemplate({
+          message: generateBotResponseTemplate({
             parsedMessage: parsedMessage,
             noflow: flow,
             flowExist: !flow,
@@ -299,8 +303,8 @@ const Flow = class extends Command {
         }
       });
 
-    if (this.getCommand().responseType || _.get(data, 'responseType')) {
-      if (_.get(data, 'responseType') === 'rich') {
+    if (this.getCommand().responseType || get(data, 'responseType')) {
+      if (get(data, 'responseType') === 'rich') {
         postMessage(
           {
             botToken: this.getBotConfig().botToken,
@@ -308,13 +312,13 @@ const Flow = class extends Command {
           },
           Object.assign(data.response, {
             channel: parsedMessage.channel,
-            thread_ts: _.get(data, 'thread')
+            thread_ts: get(data, 'thread')
               ? parsedMessage.thread_ts || parsedMessage.ts
               : '',
           })
         );
       } else {
-        responseHandler.processFile(
+        processFile(
           {
             channels: [parsedMessage.channel],
             message: {
@@ -328,7 +332,7 @@ const Flow = class extends Command {
           this.getHttpAgent()
         );
       }
-    } else if (data && _.isFunction(this.template)) {
+    } else if (data && isFunction(this.template)) {
       try {
         renderedData = this.template(data);
 
@@ -346,7 +350,7 @@ const Flow = class extends Command {
       }
     }
 
-    if (_.get(data, 'finished') === true) {
+    if (get(data, 'finished') === true) {
       this.invalidateFlow(this.questionKey, this.timerKey);
     }
 
@@ -374,11 +378,11 @@ const Flow = class extends Command {
    * @override
    */
   buildOptions(slackResponse, slackData, purpose) {
-    const user = _.find(slackData.members, {
+    const user = find(slackData.members, {
       id: slackResponse.user,
     });
 
-    const validation = _.get(this.getEventStore().get(), this.questionKey, []);
+    const validation = get(this.getEventStore().get(), this.questionKey, []);
     let response = {};
     if (validation) {
       response = validation.reduce((acc, item, index) => {
@@ -400,7 +404,7 @@ const Flow = class extends Command {
     return Object.assign(
       {
         channel: slackResponse.channel,
-        hookUrl: _.get(purpose, 'url', ''),
+        hookUrl: get(purpose, 'url', ''),
         user: user || { id: slackResponse.user },
       },
       { response }
@@ -455,8 +459,9 @@ const Flow = class extends Command {
    *
    */
   invalidateFlow(questionKey, timerKey) {
-    const timer = this.getEventStore().get(timerKey);
-    clearTimeout(timer);
+    clearTimeout(this.getEventStore().get(timerKey));
+    this.questionKey = null;
+    this.timerKey = null;
     this.getEventStore().set(timerKey, undefined);
     this.getEventStore().set(questionKey, []);
   }
@@ -496,7 +501,7 @@ const Flow = class extends Command {
 
     const alreadyAnswered = validations.filter((item) => item.answered);
     const pickCount = alreadyAnswered.length > 0 ? 3 : 2;
-    return _.take(
+    return take(
       validations.filter((item) => {
         return item.redo || !item.answered || item.preset;
       }),
@@ -505,6 +510,4 @@ const Flow = class extends Command {
   }
 };
 
-module.exports = {
-  Flow,
-};
+export { Flow };
