@@ -1,12 +1,12 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
+import * as lodash from 'lodash-es';
 
-const logger = require('../../utils/logger');
-const { Command } = require('../command');
-const responseHandler = require('../../bot/response-handler');
-
+import logger from '../../utils/logger.js';
+import { Command } from '../command.js';
+import { generateBotResponseTemplate } from '../../bot/response-handler.js';
+const { toLower, get, set, unset, isEmpty } = lodash;
 /**
  *
  * Represents the state and events of a kill command.
@@ -41,7 +41,7 @@ const Kill = class extends Command {
    */
   respond(parsedMessage) {
     const killTask = this.getParams(parsedMessage, 0);
-    const scheduleId = _.toLower(this.getParams(parsedMessage, 1));
+    const scheduleId = toLower(this.getParams(parsedMessage, 1));
     const eventStore = this.getEventStore().get();
 
     const recursiveTaskTimer = [
@@ -52,22 +52,22 @@ const Kill = class extends Command {
     const alertTimerTaskPath = [killTask, 'timer'];
     const scheduleTaskPath = [scheduleId, 'timer'];
 
-    const commandType = _.get(eventStore, [
+    const commandType = get(eventStore, [
       killTask,
       'tasks',
       scheduleId,
       'commandType',
     ]);
-    const recursiveTimer = _.get(eventStore, recursiveTaskTimer);
-    const scheduleTimer = _.get(eventStore, scheduleTaskPath);
-    const alertTimer = _.get(eventStore, alertTimerTaskPath);
+    const recursiveTimer = get(eventStore, recursiveTaskTimer);
+    const scheduleTimer = get(eventStore, scheduleTaskPath);
+    const alertTimer = get(eventStore, alertTimerTaskPath);
 
-    const targetSchedule = _.get(this.getEventStore().getSchedules(), [
+    const targetSchedule = get(this.getEventStore().getSchedules(), [
       scheduleId,
       'parsedMessage',
       'channel',
     ]);
-    const scheduledUser = _.get(this.getEventStore().getSchedules(), [
+    const scheduledUser = get(this.getEventStore().getSchedules(), [
       scheduleId,
       'parsedMessage',
       'user',
@@ -76,11 +76,11 @@ const Kill = class extends Command {
 
     if (recursiveTimer) {
       clearInterval(recursiveTimer);
-      _.set(eventStore, recursiveTaskTimer, undefined);
+      set(eventStore, recursiveTaskTimer, undefined);
 
       this.messageHandler({
         channels: parsedMessage.channel,
-        message: responseHandler.generateBotResponseTemplate({
+        message: generateBotResponseTemplate({
           parsedMessage: parsedMessage,
           command: killTask,
           /* jshint ignore:start */
@@ -105,10 +105,10 @@ const Kill = class extends Command {
           logger.error('Kill: Error killing recursive task', err);
         });
     } else if (alertTimer && commandType === 'ALERT') {
-      if (!_.get(eventStore, [killTask, 'tasks', scheduleId])) {
+      if (!get(eventStore, [killTask, 'tasks', scheduleId])) {
         this.messageHandler({
           channels: parsedMessage.channel,
-          message: responseHandler.generateBotResponseTemplate({
+          message: generateBotResponseTemplate({
             parsedMessage: parsedMessage,
             command: killTask,
             /* jshint ignore:start */
@@ -124,7 +124,7 @@ const Kill = class extends Command {
 
       this.messageHandler({
         channels: parsedMessage.channel,
-        message: responseHandler.generateBotResponseTemplate({
+        message: generateBotResponseTemplate({
           parsedMessage: parsedMessage,
           command: killTask,
           /* jshint ignore:start */
@@ -134,11 +134,11 @@ const Kill = class extends Command {
         }),
       });
 
-      _.unset(eventStore, [killTask, 'tasks', scheduleId]);
+      unset(eventStore, [killTask, 'tasks', scheduleId]);
 
-      if (_.isEmpty(_.get(eventStore, [killTask, 'tasks']))) {
+      if (isEmpty(get(eventStore, [killTask, 'tasks']))) {
         clearInterval(alertTimer);
-        _.unset(eventStore, alertTimerTaskPath);
+        unset(eventStore, alertTimerTaskPath);
       }
 
       this.getEventStore()
@@ -159,7 +159,7 @@ const Kill = class extends Command {
       if (!scheduleTimer) {
         this.messageHandler({
           channels: parsedMessage.channel,
-          message: responseHandler.generateBotResponseTemplate({
+          message: generateBotResponseTemplate({
             parsedMessage: parsedMessage,
             command: killTask,
             /* jshint ignore:start */
@@ -176,7 +176,7 @@ const Kill = class extends Command {
       if (parsedMessage.channel !== targetSchedule) {
         this.messageHandler({
           channels: parsedMessage.channel,
-          message: responseHandler.generateBotResponseTemplate({
+          message: generateBotResponseTemplate({
             parsedMessage: parsedMessage,
             command: killTask,
             /* jshint ignore:start */
@@ -192,11 +192,11 @@ const Kill = class extends Command {
       }
 
       scheduleTimer.stop();
-      _.set(eventStore, scheduleTaskPath, '');
+      set(eventStore, scheduleTaskPath, '');
 
       this.messageHandler({
         channels: parsedMessage.channel,
-        message: responseHandler.generateBotResponseTemplate({
+        message: generateBotResponseTemplate({
           parsedMessage: parsedMessage,
           command: killTask,
           /* jshint ignore:start */
@@ -226,7 +226,7 @@ const Kill = class extends Command {
     } else {
       this.messageHandler({
         channels: parsedMessage.channel,
-        message: responseHandler.generateBotResponseTemplate({
+        message: generateBotResponseTemplate({
           parsedMessage: parsedMessage,
           command: killTask,
           /* jshint ignore:start */
@@ -260,6 +260,4 @@ const Kill = class extends Command {
   }
 };
 
-module.exports = {
-  Kill,
-};
+export { Kill };
