@@ -1,12 +1,16 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
+import * as lodash from 'lodash-es';
 
-const logger = require('../../utils/logger');
-const { Command } = require('../command');
-const responseHandler = require('../../bot/response-handler');
+import logger from '../../utils/logger.js';
+import { Command } from '../command.js';
+import {
+  generateBotResponseTemplate,
+  processFile,
+} from '../../bot/response-handler.js';
 
+const { isNumber, get, isFunction } = lodash;
 /**
  *
  * Represents the state and events of a recursive command.
@@ -43,7 +47,7 @@ const Recursive = class extends Command {
     return Promise.resolve({
       then: (onFulfill) => {
         let time = Number(this.getParams(parsedMessage, 'last'));
-        time = _.isNumber(time) && time > 0 ? time : 1;
+        time = isNumber(time) && time > 0 ? time : 1;
 
         if (this.getCommand().timeUnit === 'h') {
           time = time * 3600000;
@@ -125,7 +129,7 @@ const Recursive = class extends Command {
       then: (onFulfill) => {
         this.messageHandler({
           channels: parsedMessage.channel,
-          message: responseHandler.generateBotResponseTemplate({
+          message: generateBotResponseTemplate({
             parsedMessage: parsedMessage,
             /* jshint ignore:start */
             recursiveSuccess: true,
@@ -152,8 +156,8 @@ const Recursive = class extends Command {
       return;
     }
 
-    if ((data && this.getCommand().responseType) || _.get(data, 'type')) {
-      responseHandler.processFile(
+    if ((data && this.getCommand().responseType) || get(data, 'type')) {
+      processFile(
         {
           channels: [parsedMessage.channel],
           message: {
@@ -165,7 +169,7 @@ const Recursive = class extends Command {
         this.getBotConfig().botToken,
         this.getHttpAgent()
       );
-    } else if (data && _.isFunction(this.template)) {
+    } else if (data && isFunction(this.template)) {
       try {
         this.messageHandler({
           channels: [parsedMessage.channel],
@@ -183,6 +187,4 @@ const Recursive = class extends Command {
   }
 };
 
-module.exports = {
-  Recursive,
-};
+export { Recursive };
