@@ -1,15 +1,29 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
-const CronJob = require('cron').CronJob;
-const cronTimeout = require('cron').timeout;
-const shortid = require('shortid');
+import * as lodash from 'lodash-es';
+import { CronJob } from 'cron';
+import { timeout as cronTimeout } from 'cron';
+import { nanoid } from 'nanoid';
 
-const logger = require('../../utils/logger');
-const { Command } = require('../command');
-const responseHandler = require('../../bot/response-handler');
+import logger from '../../utils/logger.js';
+import { Command } from '../command.js';
+import { generateBotResponseTemplate } from '../../bot/response-handler.js';
 
+const {
+  toUpper,
+  get,
+  toLower,
+  merge,
+  isEmpty,
+  compact,
+  join,
+  trim,
+  nth,
+  forEach,
+  slice,
+  isString,
+} = lodash;
 /**
  *
  * Represents the state and events of a schedule command.
@@ -79,8 +93,8 @@ const Schedule = class extends Command {
             cronTime: getCronExpresion(parsedMessage),
             onTick: function () {
               const scheduleCommand = getCommandArguments(parsedMessage);
-              const scheduleTask = _.toUpper(
-                _.get(scheduleCommand, 'message.command')
+              const scheduleTask = toUpper(
+                get(scheduleCommand, 'message.command')
               );
               const command = this.context[scheduleTask];
               if (command) {
@@ -93,7 +107,7 @@ const Schedule = class extends Command {
           });
 
           parsedMessage.scheduleId =
-            parsedMessage.scheduleId || _.toLower(shortid.generate());
+            parsedMessage.scheduleId || toLower(nanoid(6));
           logger.debug('schduled job for ', getCronExpresion(parsedMessage));
 
           this.setTimer(parsedMessage, job);
@@ -123,7 +137,7 @@ const Schedule = class extends Command {
 
           this.messageHandler({
             channels: parsedMessage.channel,
-            message: responseHandler.generateBotResponseTemplate({
+            message: generateBotResponseTemplate({
               parsedMessage: parsedMessage,
               noScheduledEvents: noEvents,
               scheduledEvents: scheduledEvents,
@@ -141,14 +155,14 @@ const Schedule = class extends Command {
         };
 
         try {
-          this.getCommand(_.get(scheduleCommand, 'message.command')).data.apply(
+          this.getCommand(get(scheduleCommand, 'message.command')).data.apply(
             this,
             [
               {
                 command: scheduleCommand.message.command,
                 params: scheduleCommand.message.params,
               },
-              _.merge(
+              merge(
                 {},
                 this.buildOptions(
                   scheduleCommand,
@@ -197,7 +211,7 @@ const Schedule = class extends Command {
       then: (onFulfill) => {
         this.messageHandler({
           channels: parsedMessage.channel,
-          message: responseHandler.generateBotResponseTemplate({
+          message: generateBotResponseTemplate({
             parsedMessage: parsedMessage,
             scheduleSuccess: true,
             nextEvent: this.getNextScheduleTime(parsedMessage) || false,
@@ -219,7 +233,7 @@ const Schedule = class extends Command {
    */
   validate(parsedMessage) {
     const scheduleCommand = getCommandArguments(parsedMessage);
-    const scheduleTask = _.toUpper(_.get(scheduleCommand, 'message.command'));
+    const scheduleTask = toUpper(get(scheduleCommand, 'message.command'));
     const command = this.context[scheduleTask];
 
     if (scheduleTask === 'LIST') {
@@ -263,10 +277,7 @@ const Schedule = class extends Command {
     return Promise.resolve({
       then: (onFulfill, onReject) => {
         try {
-          if (
-            _.isEmpty(cron) ||
-            _.compact((cron || '').split(' ')).length > 6
-          ) {
+          if (isEmpty(cron) || compact((cron || '').split(' ')).length > 6) {
             logger.error(
               'Invalid cron ',
               cron,
@@ -317,10 +328,7 @@ const Schedule = class extends Command {
    * @override
    */
   getTimer(parsedMessage) {
-    return _.get(this.getEventStore().get(), [
-      parsedMessage.scheduleId,
-      'timer',
-    ]);
+    return get(this.getEventStore().get(), [parsedMessage.scheduleId, 'timer']);
   }
 
   /**
@@ -370,11 +378,9 @@ const Schedule = class extends Command {
  */
 const getCronExpresion = function (parsedMessage) {
   const cronRegex = /\((.*?)\)/;
-  const cronExpresion = cronRegex.exec(
-    _.join(parsedMessage.message.params, ' ')
-  );
+  const cronExpresion = cronRegex.exec(join(parsedMessage.message.params, ' '));
 
-  return _.trim('0 ' + _.nth(cronExpresion, 1));
+  return trim('0 ' + nth(cronExpresion, 1));
 };
 
 /**
@@ -385,14 +391,14 @@ const getCronExpresion = function (parsedMessage) {
  */
 const getCommandArguments = function (parsedMessage) {
   const result = [];
-  _.forEach(
-    _.slice(
-      _.get(parsedMessage, 'message.params'),
+  forEach(
+    slice(
+      get(parsedMessage, 'message.params'),
       1,
-      _.get(parsedMessage, 'message.params', []).length
+      get(parsedMessage, 'message.params', []).length
     ),
     function (value) {
-      if (_.isString(value) && value.indexOf('(') > -1) {
+      if (isString(value) && value.indexOf('(') > -1) {
         return false;
       }
       result.push(value);
@@ -404,12 +410,10 @@ const getCommandArguments = function (parsedMessage) {
     channel: parsedMessage.channel,
     user: parsedMessage.user,
     message: {
-      command: _.nth(_.get(parsedMessage, 'message.params')),
+      command: nth(get(parsedMessage, 'message.params')),
       params: result,
     },
   };
 };
 
-module.exports = {
-  Schedule,
-};
+export { Schedule };
