@@ -1,9 +1,9 @@
 'use strict';
 
-const _ = require('lodash');
+import * as lodash from 'lodash-es';
+import logger from '../utils/logger.js';
 
-const logger = require('../utils/logger');
-
+const { curry, compact, get, trim, toUpper, nth, forEach, map, slice } = lodash;
 /**
 * Function to parse incoming slack message.
 *
@@ -14,10 +14,10 @@ const logger = require('../utils/logger');
 
 * @return {object} parsed bot message.
 */
-const parse = _.curry(function (commandList, options, message) {
+const parse = curry(function (commandList, options, message) {
   const parsedCommand = Object.assign({}, message);
-  const messageArr = _.compact(
-    _.get(message, 'text', '')
+  const messageArr = compact(
+    get(message, 'text', '')
       /* eslint-disable no-control-regex */
       .replace(/[^\x00-\x7F]/g, ' ')
       .replace(/^<@(.*?)>/, (str) => `${str} `)
@@ -45,25 +45,25 @@ const parse = _.curry(function (commandList, options, message) {
 const mapCommand = function (commandList, messageArr, options) {
   const firstArg = messageArr[0];
   const secondArg = messageArr[1];
-  const upperCaseFirstArg = _.trim(_.toUpper(firstArg));
-  const upperCaseSecondArg = _.trim(_.toUpper(secondArg));
+  const upperCaseFirstArg = trim(toUpper(firstArg));
+  const upperCaseSecondArg = trim(toUpper(secondArg));
   const botMentionNameRegex = new RegExp(/^<@(.*?)>/);
   const botNameRegex = new RegExp(/^@(.*?)?$/);
-  const botId = _.nth(botMentionNameRegex.exec(firstArg), 1);
-  const botName = _.toUpper(
-    _.nth(botNameRegex.exec(firstArg), 1) || upperCaseFirstArg
+  const botId = nth(botMentionNameRegex.exec(firstArg), 1);
+  const botName = toUpper(
+    nth(botNameRegex.exec(firstArg), 1) || upperCaseFirstArg
   );
   const messageMap = {};
   const isEid = options.eId ? options.eId === botId : false;
 
-  if (options.id === botId || isEid || _.toUpper(options.name) === botName) {
+  if (options.id === botId || isEid || toUpper(options.name) === botName) {
     messageMap.commandPrefix = botId || botName;
   }
 
-  _.forEach(commandList, function (command) {
+  forEach(commandList, function (command) {
     if (messageMap.commandPrefix) {
       if (
-        _.trim(command.command) === upperCaseSecondArg ||
+        trim(command.command) === upperCaseSecondArg ||
         command.alias.indexOf(upperCaseSecondArg) > -1
       ) {
         messageMap.command = command.command;
@@ -82,13 +82,9 @@ const mapCommand = function (commandList, messageArr, options) {
 };
 
 const mapParams = function (messageArr, position) {
-  return _.map(_.slice(messageArr, position, messageArr.length), function (
-    item
-  ) {
+  return map(slice(messageArr, position, messageArr.length), function (item) {
     return item;
   });
 };
 
-module.exports = {
-  parse,
-};
+export { parse };
