@@ -1,16 +1,16 @@
 'use strict';
 
 // Load modules
-const _ = require('lodash');
-const url = require('url');
-const assert = require('assert');
+import * as lodash from 'lodash-es';
+import { parse } from 'url';
+import { ok } from 'assert';
 
-const logger = require('./utils/logger');
-const { Bots } = require('./bot/bots');
-const socketServer = require('./bot/socket-server');
-const socket = require('./bot/socket');
-const server = require('./bot/server');
-const storage = require('./storage/storage');
+import logger from './utils/logger.js';
+import { Bots } from './bot/bots.js';
+import { connect } from './bot/socket-server.js';
+import * as Socket from './bot/socket.js';
+import { setupServer as _setupServer } from './bot/server.js';
+import { Storage } from './storage/storage.js';
 
 const DEFAULT_CONFIG = {
   defaultConfig: {
@@ -22,6 +22,17 @@ const DEFAULT_CONFIG = {
     },
   },
 };
+const {
+  cloneDeep,
+  get,
+  split,
+  nth,
+  isEmpty,
+  forEach,
+  includes,
+  toUpper,
+  isUndefined,
+} = lodash;
 
 /**
  *
@@ -39,13 +50,13 @@ const SlackBot = class {
   constructor(options, settings) {
     this.config = Object.assign(
       DEFAULT_CONFIG.defaultConfig.slackBotRoot,
-      _.cloneDeep(options)
+      cloneDeep(options)
     );
     this.settings = settings;
 
     this.assertInputData(this.config, settings);
 
-    storage.createEventDirectory();
+    Storage.createEventDirectory();
     logger.setLogger(this.config.logger);
 
     this.bots = new Bots(this.config).getBots();
@@ -66,9 +77,8 @@ const SlackBot = class {
           logger.error('Failed setting up webhook server %j', err);
         });
     } else {
-      if (_.get(this.settings, 'isMock')) {
-        return socketServer
-          .connect()
+      if (get(this.settings, 'isMock')) {
+        return connect()
           .then(() => {
             return this.initializeBots(null);
           })
@@ -100,7 +110,7 @@ const SlackBot = class {
    */
   setupServer(config) {
     if (config.server) {
-      return server.setupServer(config.server, (request, response) => {
+      return _setupServer(config.server, (request, response) => {
         this.handleRoutes.call(this, request, response);
       });
     }
@@ -112,7 +122,7 @@ const SlackBot = class {
    * @param {object} response http response object.
    */
   handleRoutes(request, response) {
-    const urlParts = url.parse(request.url);
+    const urlParts = parse(request.url);
     let body = '';
 
     request.on('data', (chunk) => {
@@ -140,17 +150,17 @@ const SlackBot = class {
    * @param {object} response http response object.
    */
   selectRoutes(route, requestBody, request, response) {
-    const paths = _.split(route.pathname, '/');
-    const appRoute = _.nth(paths, 1);
+    const paths = split(route.pathname, '/');
+    const appRoute = nth(paths, 1);
 
     if (
       appRoute === 'hook' &&
       request.method === 'POST' &&
-      !_.isEmpty(this.bots)
+      !isEmpty(this.bots)
     ) {
-      const botId = _.nth(paths, 2);
-      const purposeId = _.nth(paths, 3);
-      _.forEach(this.bots, (bot) => {
+      const botId = nth(paths, 2);
+      const purposeId = nth(paths, 3);
+      forEach(this.bots, (bot) => {
         if (botId === bot.getId()) {
           bot.handleHookRequest(purposeId, requestBody, response);
         }
@@ -214,7 +224,7 @@ const SlackBot = class {
   closeSocketSession(botInstances, bot) {
     return botInstances
       .then(function () {
-        socket.closeConnection(bot);
+        Socket.closeConnection(bot);
       })
       .catch(function (err) {
         logger.error('Closing bot failed ', err);
@@ -245,31 +255,31 @@ const SlackBot = class {
    * @param {object} settings holds mock settings.
    */
   assertInputData(config, settings) {
-    assert.ok(config.bots, 'Invalid or empty config passed. Refer link here');
-    assert.ok(config.bots.length > 0, 'Bots cannot be empty. Refer link here');
-    _.forEach(config.bots, function (bot) {
-      if (!_.get(settings, 'isMock')) {
+    ok(config.bots, 'Invalid or empty config passed. Refer link here');
+    ok(config.bots.length > 0, 'Bots cannot be empty. Refer link here');
+    forEach(config.bots, function (bot) {
+      if (!get(settings, 'isMock')) {
         delete bot.mock;
       }
 
-      assert.ok(
-        !_.isEmpty(bot.botToken),
+      ok(
+        !isEmpty(bot.botToken),
         'Bot need to have bot token. Refer github docs.'
       );
 
-      assert.ok(
-        !_.isEmpty(bot.botCommand),
+      ok(
+        !isEmpty(bot.botCommand),
         'Bot need to have atleast one command. Refer github docs.'
       );
 
-      assert.ok(
-        !_.isEmpty(bot.botCommand),
+      ok(
+        !isEmpty(bot.botCommand),
         'Bot need to have atleast one command. Refer github docs.'
       );
 
-      _.forEach(bot.botCommand, function (botCommand, key) {
-        assert.ok(
-          !_.isEmpty(botCommand.commandType),
+      forEach(bot.botCommand, function (botCommand, key) {
+        ok(
+          !isEmpty(botCommand.commandType),
           'Each bot should have command type. Bot: ' +
             bot.name +
             ' Key: ' +
@@ -277,18 +287,18 @@ const SlackBot = class {
             ' Refer github docs.'
         );
 
-        assert.ok(
-          _.includes(
+        ok(
+          includes(
             ['DATA', 'RECURSIVE', 'KILL', 'ALERT', 'FLOW'],
-            _.toUpper(botCommand.commandType)
+            toUpper(botCommand.commandType)
           ),
           'Unrecognized bot command type. Only "data",' +
             ' "recursive", "alert", "kill" are supported'
         );
 
-        if (_.includes(['ALERT'], _.toUpper(botCommand.commandType))) {
-          assert.ok(
-            !_.isUndefined(botCommand.timeInterval),
+        if (includes(['ALERT'], toUpper(botCommand.commandType))) {
+          ok(
+            !isUndefined(botCommand.timeInterval),
             'Bot command of type "alert" should have timeInterval'
           );
         }
@@ -297,4 +307,4 @@ const SlackBot = class {
   }
 };
 
-module.exports = SlackBot;
+export default SlackBot;
