@@ -1,8 +1,10 @@
 'use strict';
 
-const _ = require('lodash');
-const http = require('https');
-const logger = require('../utils/logger');
+import * as lodash from 'lodash-es';
+import { request } from 'https';
+import logger from '../utils/logger.js';
+
+const { extend, get, nth, split, clone, partial } = lodash;
 
 /**
  * Function to make slack api get request.
@@ -19,7 +21,7 @@ const fetch = function (options, offset) {
 
   return Promise.resolve({
     then: (onFulfill, onReject) => {
-      const httpOptions = _.extend(
+      const httpOptions = extend(
         {
           host: 'slack.com',
           method: 'GET',
@@ -31,56 +33,54 @@ const fetch = function (options, offset) {
         options
       );
 
-      const req = http
-        .request(httpOptions, function (response) {
-          let responseStr = '';
+      const req = request(httpOptions, function (response) {
+        let responseStr = '';
 
-          response.on('data', function (chunk) {
-            responseStr += chunk;
-          });
-
-          response.on('end', function () {
-            let slackResponse = {};
-            let error = '';
-
-            try {
-              slackResponse = JSON.parse(responseStr);
-            } catch (err) {
-              logger.error('Slack response corrupted ', responseStr);
-              slackResponse.error = 'invalid_slack_response';
-            }
-
-            error = _.get(slackResponse, 'error');
-
-            if (error === 'invalid_auth' || error === 'not_authed') {
-              logger.info('Invalid auth token. Unable to connect');
-              slackResponse.error = error;
-
-              return onReject(slackResponse);
-            } else if (error) {
-              slackResponse.error = _.get(slackResponse, 'error');
-              logger.info('Something wrong ', slackResponse.error);
-              logger.debug('Something wrong with request for ', httpOptions);
-
-              return onReject(slackResponse);
-            } else {
-              logger.info(
-                'Api call returned success for ',
-                _.nth(_.split(httpOptions.path, '?', 1), 0),
-                offset
-              );
-              logger.debug('Api call response success ', slackResponse);
-
-              return onFulfill(slackResponse);
-            }
-          });
-        })
-        .on('error', function (err) {
-          logger.info('Unknown error on connecting to slack ', err);
-          onReject({
-            error: 'unkown_error',
-          });
+        response.on('data', function (chunk) {
+          responseStr += chunk;
         });
+
+        response.on('end', function () {
+          let slackResponse = {};
+          let errorRes = '';
+
+          try {
+            slackResponse = JSON.parse(responseStr);
+          } catch (err) {
+            logger.error('Slack response corrupted ', responseStr);
+            slackResponse.error = 'invalid_slack_response';
+          }
+
+          errorRes = get(slackResponse, 'error');
+
+          if (errorRes === 'invalid_auth' || errorRes === 'not_authed') {
+            logger.info('Invalid auth token. Unable to connect');
+            slackResponse.error = errorRes;
+
+            return onReject(slackResponse);
+          } else if (errorRes) {
+            slackResponse.error = get(slackResponse, 'error');
+            logger.info('Something wrong ', slackResponse.error);
+            logger.debug('Something wrong with request for ', httpOptions);
+
+            return onReject(slackResponse);
+          } else {
+            logger.info(
+              'Api call returned success for ',
+              nth(split(httpOptions.path, '?', 1), 0),
+              offset
+            );
+            logger.debug('Api call response success ', slackResponse);
+
+            return onFulfill(slackResponse);
+          }
+        });
+      }).on('error', function (err) {
+        logger.info('Unknown error on connecting to slack ', err);
+        onReject({
+          error: 'unkown_error',
+        });
+      });
       req.end();
     },
   });
@@ -92,14 +92,14 @@ const fetchBatchWithRetry = function (options, interval, handler) {
   let offset;
 
   const initiateRequest = (offset, onFulfill, onReject) => {
-    const apiOptions = _.clone(options);
+    const apiOptions = clone(options);
 
     return fetch(apiOptions, offset)
       .then((response) => {
-        offset = _.get(response, 'offset');
+        offset = get(response, 'offset');
         store = handler(response, store);
 
-        if (_.get(response, 'response_metadata.next_cursor')) {
+        if (get(response, 'response_metadata.next_cursor')) {
           return timedRetryHandler(
             initiateRequest,
             offset,
@@ -135,7 +135,7 @@ const fetchBatchWithRetry = function (options, interval, handler) {
   };
 
   return Promise.resolve({
-    then: _.partial((onFulfill, onReject) => {
+    then: partial((onFulfill, onReject) => {
       return initiateRequest(offset, onFulfill, onReject);
     }),
   });
@@ -152,7 +152,7 @@ const fetchBatchWithRetry = function (options, interval, handler) {
 const post = function (options, data) {
   return Promise.resolve({
     then: (onFulfill, onReject) => {
-      const httpOptions = _.extend(
+      const httpOptions = extend(
         {
           host: 'slack.com',
           method: 'POST',
@@ -165,57 +165,55 @@ const post = function (options, data) {
         options
       );
 
-      const req = http
-        .request(httpOptions, function (response) {
-          let responseStr = '';
+      const req = request(httpOptions, function (response) {
+        let responseStr = '';
 
-          response.on('data', function (chunk) {
-            responseStr += chunk;
-          });
-
-          response.on('end', function () {
-            let slackResponse = {};
-            let error = '';
-            try {
-              slackResponse = JSON.parse(responseStr);
-            } catch (err) {
-              logger.error('Slack response corrupted ', responseStr);
-              slackResponse.error = 'invalid_slack_response';
-            }
-
-            error = _.get(slackResponse, 'error');
-
-            if (error === 'invalid_auth' || error === 'not_authed') {
-              logger.info('Invalid auth token. Unable to connect');
-              slackResponse.error = error;
-
-              return onReject(slackResponse);
-            } else if (error) {
-              slackResponse.error = _.get(slackResponse, 'error');
-              logger.info('Something wrong ', slackResponse.error);
-              logger.debug('Something wrong with request for ', httpOptions);
-
-              return onReject(slackResponse);
-            } else {
-              logger.debug('Api call response success ', slackResponse);
-
-              return onFulfill(slackResponse);
-            }
-          });
-        })
-        .on('error', function (err) {
-          logger.info('Unknown error on connecting to slack ', err);
-          onReject({
-            error: 'unkown_error',
-          });
+        response.on('data', function (chunk) {
+          responseStr += chunk;
         });
+
+        response.on('end', function () {
+          let slackResponse = {};
+          let error = '';
+          try {
+            slackResponse = JSON.parse(responseStr);
+          } catch (err) {
+            logger.error('Slack response corrupted ', responseStr);
+            slackResponse.error = 'invalid_slack_response';
+          }
+
+          error = get(slackResponse, 'error');
+
+          if (error === 'invalid_auth' || error === 'not_authed') {
+            logger.info('Invalid auth token. Unable to connect');
+            slackResponse.error = error;
+
+            return onReject(slackResponse);
+          } else if (error) {
+            slackResponse.error = get(slackResponse, 'error');
+            logger.info('Something wrong ', slackResponse.error);
+            logger.debug('Something wrong with request for ', httpOptions);
+
+            return onReject(slackResponse);
+          } else {
+            logger.debug('Api call response success ', slackResponse);
+
+            return onFulfill(slackResponse);
+          }
+        });
+      }).on('error', function (err) {
+        logger.info('Unknown error on connecting to slack ', err);
+        onReject({
+          error: 'unkown_error',
+        });
+      });
       req.write(JSON.stringify(data));
       req.end();
     },
   });
 };
 
-module.exports = {
+export default {
   fetch,
   fetchBatchWithRetry,
   post,
