@@ -1,7 +1,9 @@
 'use strict';
 
-const _ = require('lodash');
-const apiRequest = require('./api-request');
+import * as lodash from 'lodash-es';
+import * as apiRequest from './api-request.js';
+
+const { concat, get } = lodash;
 
 /**
  * Function to make get channel request.
@@ -14,14 +16,14 @@ const getChannelsList = function (options) {
 };
 
 const resultHandler = function (response, channels = []) {
-  return _.concat(channels, _.get(response, 'channels'));
+  return concat(channels, get(response, 'channels'));
 };
 
 const requestChannelList = function (options, handler) {
-  const excludeMembers = _.get(options, 'excludeMembers', 1);
-  const excludeArchived = _.get(options, 'excludeArchived', 1);
-  const limit = _.get(options, 'limit', 999);
-  const interval = _.get(options, 'interval');
+  const excludeMembers = get(options, 'excludeMembers', 1);
+  const excludeArchived = get(options, 'excludeArchived', 1);
+  const limit = get(options, 'limit', 999);
+  const interval = get(options, 'interval');
 
   const path =
     '/api/channels.list?' +
@@ -44,6 +46,4 @@ const requestChannelList = function (options, handler) {
   );
 };
 
-module.exports = {
-  getChannelsList,
-};
+export { getChannelsList };
