@@ -1,8 +1,8 @@
 'use strict';
 
-const http = require('https');
-const FormDataHandler = require('form-data');
-const mime = require('mime');
+import { request as _request } from 'https';
+import FormDataHandler from 'form-data';
+import * as Mime from 'mime';
 
 /**
  * Function to handle file uploads
@@ -22,8 +22,8 @@ const postAttachmentFromStream = function (channel, config, data) {
       form.append('channels', channel[0]);
       form.append('file', data, {
         filename: config.name || config.commandName + '_graph.' + fileType,
-        contentType: mime.getType(fileType),
-        filetype: mime.getType(fileType),
+        contentType: Mime.getType(fileType),
+        filetype: Mime.getType(fileType),
       });
 
       const options = {
@@ -42,7 +42,7 @@ const postAttachmentFromStream = function (channel, config, data) {
         options.path = options.path + '?thread_ts=' + config.thread;
       }
 
-      const request = http.request(options, (res) => {
+      const request = _request(options, (res) => {
         let responseStr = '';
 
         res.on('data', function (chunk) {
@@ -74,6 +74,4 @@ const postAttachmentFromStream = function (channel, config, data) {
   });
 };
 
-module.exports = {
-  postAttachmentFromStream,
-};
+export { postAttachmentFromStream };
