@@ -1,6 +1,6 @@
 'use strict';
 
-const apiRequest = require('./api-request');
+import apiRequest from './api-request.js';
 
 /**
  * Function to make post message to slack.
@@ -21,4 +21,4 @@ const postMessage = function (options, data) {
   );
 };
 
-module.exports = postMessage;
+export default postMessage;
