@@ -1,6 +1,6 @@
 'use strict';
 
-const apiRequest = require('./api-request');
+import apiRequest from './api-request.js';
 
 /**
  * Function to make get users group request.
@@ -15,6 +15,4 @@ const getUserGroupsList = function (options) {
   });
 };
 
-module.exports = {
-  getUserGroupsList,
-};
+export { getUserGroupsList };
