@@ -1,7 +1,9 @@
 'use strict';
 
-const _ = require('lodash');
-const apiRequest = require('./api-request');
+import apiRequest from './api-request.js';
+import * as lodash from 'lodash-es';
+
+const { concat, get } = lodash;
 
 /**
  * Function to make get users request.
@@ -14,15 +16,15 @@ const getUsersList = function (options) {
 };
 
 const resultHandler = function (response, members = []) {
-  members = _.concat(members, _.get(response, 'members'));
+  members = concat(members, get(response, 'members'));
 
   return members;
 };
 
 const requestUserList = function (options, handler) {
-  const presence = _.get(options, 'presence', 0);
-  const limit = _.get(options, 'limit', 0);
-  const interval = _.get(options, 'interval');
+  const presence = get(options, 'presence', 0);
+  const limit = get(options, 'limit', 0);
+  const interval = get(options, 'interval');
 
   const path =
     '/api/users.list?' +
@@ -44,7 +46,7 @@ const requestUserList = function (options, handler) {
 };
 
 const getUserInfo = function (options) {
-  const user = _.get(options, 'userId');
+  const user = get(options, 'userId');
 
   const path =
     '/api/users.info?' + 'user=' + user + '&token=' + options.botToken;
@@ -55,7 +57,4 @@ const getUserInfo = function (options) {
   });
 };
 
-module.exports = {
-  getUsersList,
-  getUserInfo,
-};
+export { getUsersList, getUserInfo };
