@@ -1,15 +1,16 @@
 'use strict';
 
-const fs = require('fs');
-const _ = require('lodash');
-const path = require('path');
+import { mkdir, readFile as _readFile, writeFile as _writeFile } from 'fs';
+import * as lodash from 'lodash-es';
+import { join } from 'path';
 
-const logger = require('../utils/logger');
+import logger from '../utils/logger.js';
 
-const STORAGE_DIRECTORY = path.join(process.cwd(), 'data');
-const EVENT_FILE_PATH = path.join(STORAGE_DIRECTORY, 'events.json');
-const SCHEDULE_FILE_PATH = path.join(STORAGE_DIRECTORY, 'schedule.json');
+const STORAGE_DIRECTORY = join(process.cwd(), 'data');
+const EVENT_FILE_PATH = join(STORAGE_DIRECTORY, 'events.json');
+const SCHEDULE_FILE_PATH = join(STORAGE_DIRECTORY, 'schedule.json');
 
+const { set, get, forEach, concat, toLower, unset, toUpper } = lodash;
 const Storage = class {
   /**
    * Updates events.
@@ -24,10 +25,10 @@ const Storage = class {
       .then((eventsData) => {
         if (data && data.parsedMessage && data.channels) {
           const result = eventsData || {};
-          _.set(
+          set(
             result,
             options.botName,
-            pickEvents(options, _.get(eventsData, options.botName, {}), data)
+            pickEvents(options, get(eventsData, options.botName, {}), data)
           );
 
           return this.writeFile(options.eventType, result);
@@ -60,19 +61,19 @@ const Storage = class {
       then: (onFulfill, onReject) => {
         this.readFile(options.eventType)
           .then((eventsData) => {
-            if (_.get(data, 'channels', []).length) {
-              _.forEach(data.channels, function () {
-                const eventPath = _.concat(
+            if (get(data, 'channels', []).length) {
+              forEach(data.channels, function () {
+                const eventPath = concat(
                   options.botName,
-                  _.get(data, 'commandToKill')
+                  get(data, 'commandToKill')
                 );
                 const compatibleEventPath = [
                   options.botName,
-                  _.toLower(_.get(data, 'commandToKill')),
+                  toLower(get(data, 'commandToKill')),
                 ];
 
-                if (_.unset(eventsData, eventPath)) {
-                  _.unset(eventsData, compatibleEventPath);
+                if (unset(eventsData, eventPath)) {
+                  unset(eventsData, compatibleEventPath);
                   logger.info('Events updates successfully');
                 }
               });
@@ -120,7 +121,7 @@ const Storage = class {
    * Get directory for events.
    */
   static createEventDirectory() {
-    fs.mkdir(STORAGE_DIRECTORY, function (e) {
+    mkdir(STORAGE_DIRECTORY, function (e) {
       if (!e || (e && e.code === 'EEXIST')) {
         logger.info('Data directory already exist, so not creating');
       } else {
@@ -151,7 +152,7 @@ const Storage = class {
 
     return Promise.resolve({
       then: (onFulfill, onReject) => {
-        fs.readFile(
+        _readFile(
           path,
           {
             encoding: 'utf8',
@@ -196,7 +197,7 @@ const Storage = class {
 
     return Promise.resolve({
       then: (onFulfill, onReject) => {
-        fs.writeFile(
+        _writeFile(
           path,
           fileData,
           {
@@ -218,17 +219,17 @@ const Storage = class {
 };
 
 const pickEvents = function (options, storeData, newData) {
-  const eventTask = _.toUpper(_.get(newData, 'parsedMessage.message.command'));
+  const eventTask = toUpper(get(newData, 'parsedMessage.message.command'));
 
-  _.forEach(newData.channels, function (channel) {
+  forEach(newData.channels, function (channel) {
     if (options.id) {
-      _.set(storeData, options.id, newData);
+      set(storeData, options.id, newData);
     } else {
-      _.set(storeData, channel + '_' + eventTask, newData);
+      set(storeData, channel + '_' + eventTask, newData);
     }
   });
 
   return storeData;
 };
 
-module.exports = Storage;
+export { Storage };
