@@ -1,7 +1,9 @@
 'use strict';
 
-const _ = require('lodash');
-const env = require('../utils/environment')();
+import * as lodash from 'lodash-es';
+import env from '../utils/environment.js';
+
+const { isFunction, compact, trim, get } = lodash;
 
 /**
  * Represents set and methods for logging.
@@ -12,8 +14,21 @@ const Logger = class {
    *
    * @class
    */
-  constructor() {
-    this.logger = console;
+  constructor(logger) {
+    if (logger) {
+      this.setLog(logger);
+    }
+  }
+
+  /**
+   * Function to log info level.
+   *
+   * @param {string} message message to log.
+   * @param {array} args argument to logger handle.
+   */
+  log(message, ...args) {
+    args = args ? args : '';
+    this._log(this.getLog().log, 'log', message, ...args);
   }
 
   /**
@@ -24,7 +39,7 @@ const Logger = class {
    */
   info(message, ...args) {
     args = args ? args : '';
-    this.log(this.logger.info, 'info', message, ...args);
+    this._log(this.getLog().log, 'info', message, ...args);
   }
 
   /**
@@ -35,7 +50,7 @@ const Logger = class {
    */
   debug(message, ...args) {
     args = args ? args : '';
-    this.log(this.logger.debug, 'debug', message, ...args);
+    this._log(this.getLog().debug, 'debug', message, ...args);
   }
 
   /**
@@ -46,7 +61,7 @@ const Logger = class {
    */
   error(message, ...args) {
     args = args ? args : '';
-    this.log(this.logger.error, 'error', message, ...args);
+    this._log(this.getLog().error, 'error', message, ...args);
   }
 
   /**
@@ -57,19 +72,17 @@ const Logger = class {
    * @param {string} message log message.
    * @param {array} args log arguments.
    */
-  log(handler, level, message, ...args) {
+  _log(handler, level, message, ...args) {
+    const logHandler = isFunction(handler) ? handler : this.getLog().log;
+
     if (this.isTest()) {
       return;
-    } else if (
-      this.isConsole() &&
-      !this.isDebug() &&
-      (level === 'info' || level === 'error')
-    ) {
-      this.logger.log(this.prefixTime(message, ...args));
-    } else if (this.isConsole() && this.isDebug()) {
-      this.logger.log(this.prefixTime(message, ...args));
-    } else if (_.isFunction(handler)) {
-      handler(message, ...args);
+    } else if (this.isProd(level)) {
+      logHandler('%j', this.prefixTime(message, ...args));
+    } else if (this.isNonProd(level)) {
+      logHandler('%j', this.prefixTime(message, ...args));
+    } else if (this.isDebug(level)) {
+      logHandler('%j', this.prefixTime(message, ...args));
     }
   }
 
@@ -79,7 +92,7 @@ const Logger = class {
    * @param {handler} logger logger handler.
    */
   setLogger(logger) {
-    this.logger = logger ? logger : console;
+    return this.setLog(logger || console);
   }
 
   /**
@@ -96,12 +109,12 @@ const Logger = class {
     if (message) {
       log = {
         time: time,
-        message: _.trim(message),
+        message: trim(message),
       };
     }
 
     if ((args || []).length > 0) {
-      log.data = _.compact(args);
+      log.data = compact(args);
     }
 
     return log;
@@ -112,8 +125,26 @@ const Logger = class {
    *
    * @return {boolean} true or false.
    */
-  isDebug() {
-    return !env.test && env.dev;
+  isDebug(level) {
+    return this.getEnv().dev && level == 'debug';
+  }
+
+  /**
+   * Function to validate if NODE_ENV is prod mode.
+   *
+   * @return {boolean} true or false.
+   */
+  isProd(level) {
+    return this.getEnv().prod && level !== 'debug';
+  }
+
+  /**
+   * Function to validate if NODE_ENV is prod mode.
+   *
+   * @return {boolean} true or false.
+   */
+  isNonProd() {
+    return !this.getEnv().prod;
   }
 
   /**
@@ -122,17 +153,35 @@ const Logger = class {
    * @return {boolean} true or false.
    */
   isTest() {
-    return !!env.test;
+    return !!this.getEnv().test;
   }
 
+  /**
+   * Function to get env.
+   *
+   * @return {object} prod or test or dev.
+   */
+  getEnv() {
+    return env;
+  }
   /**
    * Function to validate if log handler is console.
    *
    * @return {boolean} true or false.
    */
   isConsole() {
-    return _.get(this, 'logger.constructor.name') === 'Console';
+    return get(this, 'logger.constructor.name') === 'Console';
+  }
+
+  setLog(logger) {
+    this.logger = logger;
+    return logger;
+  }
+
+  getLog() {
+    return this.logger;
   }
 };
 
-module.exports = new Logger();
+const loggerInstance = new Logger(console);
+export default loggerInstance;
