@@ -1,15 +1,15 @@
 'use strict';
 
-const _ = require('lodash');
-const { Bots } = require('./../../../lib/bot/bots');
-const config = require('../../mock');
+import { forEach } from 'lodash-es';
+import { Bots } from './../../../lib/bot/bots.js';
+import fixtures from '../../mock/index.js';
 
 describe('/bots', function () {
   describe('Should instantiate bots correctly', function () {
     let slackBots;
 
     beforeEach(function () {
-      slackBots = new Bots(config.BotsTest.bots).getBots();
+      slackBots = new Bots(fixtures.BotsTest.bots).getBots();
     });
 
     afterEach(function () {
@@ -19,14 +19,14 @@ describe('/bots', function () {
     describe('Should instantiate bots correctly', function () {
       it('Should contain bot token and command for bots', function () {
         expect(slackBots).toBeTruthy();
-        _.forEach(slackBots, function (botInfo) {
+        forEach(slackBots, function (botInfo) {
           expect(botInfo.config.botCommand).toBeTruthy();
         });
       });
 
       it('Should contain normalized bots', function () {
         expect(slackBots).toBeTruthy();
-        _.forEach(slackBots, function (botInfo) {
+        forEach(slackBots, function (botInfo) {
           expect(botInfo.config.botCommand['PING-ME']).toBeTruthy();
           expect(botInfo.config.botCommand['STOP']).toBeUndefined();
         });
@@ -38,7 +38,7 @@ describe('/bots', function () {
     let slackBots;
 
     beforeEach(function () {
-      slackBots = new Bots(config.BotsTestWithRecursiveTasks.bots).getBots();
+      slackBots = new Bots(fixtures.BotsTestWithRecursiveTasks.bots).getBots();
     });
 
     afterEach(function () {
@@ -48,14 +48,14 @@ describe('/bots', function () {
     describe('Should instantiate bots correctly', function () {
       it('Should contain bot token and command for bots', function () {
         expect(slackBots).toBeTruthy();
-        _.forEach(slackBots, function (botInfo) {
+        forEach(slackBots, function (botInfo) {
           expect(botInfo.config.botCommand).toBeTruthy();
         });
       });
 
       it('Should contain normalized bots', function () {
         expect(slackBots).toBeTruthy();
-        _.forEach(slackBots, function (botInfo) {
+        forEach(slackBots, function (botInfo) {
           expect(botInfo.config.botCommand['PING-ME']).toBeTruthy();
           expect(botInfo.config.botCommand['AUTODATA']).toBeTruthy();
           expect(botInfo.config.botCommand['STOP']).toBeTruthy();
