@@ -5,7 +5,8 @@
  *
  * @return {object} environment set in NODE_ENV.
  */
-module.exports = function () {
+
+const env = (function () {
   if (process.env.NODE_ENV) {
     return {
       [process.env.NODE_ENV]: true,
@@ -15,4 +16,6 @@ module.exports = function () {
   return {
     prod: true,
   };
-};
+})();
+
+export default env;
